Split compiler code into separate files.
