+++ src/twig.exports.compile.js
@@ -1,0 +1,81 @@
+// ## twig.exports.compile.js
+//
+// This file provides extension points and other hooks into twig compilation.
+
+module.exports = function (Twig) {
+    "use strict";
+
+    /**
+     * Provide an extension for use with express 2.
+     *
+     * @param {string} markup The template markup.
+     * @param {array} options The express options.
+     *
+     * @return {string} The rendered template.
+     */
+    Twig.exports.compile = function(markup, options) {
+        var id = options.filename,
+            path = options.filename,
+            template;
+
+        // Try to load the template from the cache
+        template = new Twig.Template({
+            data: markup,
+            path: path,
+            id: id,
+            options: options.settings['twig options']
+        }); // Twig.Templates.load(id) ||
+
+        return function(context) {
+            return template.render(context);
+        };
+    };
+
+    /**
+     * Provide an extension for use with express 3.
+     *
+     * @param {string} path The location of the template file on disk.
+     * @param {Object|Function} The options or callback.
+     * @param {Function} fn callback.
+     *
+     * @throws Twig.Error
+     */
+    Twig.exports.renderFile = function(path, options, fn) {
+        // handle callback in options
+        if (typeof options === 'function') {
+            fn = options;
+            options = {};
+        }
+
+        options = options || {};
+
+        var settings = options.settings || {};
+
+        var params = {
+            path: path,
+            base: settings.views,
+            load: function(template) {
+                // render and return template as a simple string, see https://github.com/twigjs/twig.js/pull/348 for more information
+                fn(null, '' + template.render(options));
+            }
+        };
+
+        // mixin any options provided to the express app.
+        var view_options = settings['twig options'];
+
+        if (view_options) {
+            for (var option in view_options) {
+                if (view_options.hasOwnProperty(option)) {
+                    params[option] = view_options[option];
+                }
+            }
+        }
+
+        Twig.exports.twig(params);
+    };
+
+    // Express 3 handler
+    Twig.exports.__express = Twig.exports.renderFile;
+
+    return Twig;
+};
