+++ src/twig.expression.compile.js
@@ -1,0 +1,161 @@
+// ## twig.expression.compile.js
+//
+// This file handles compiling expressions.
+module.exports = function (Twig) {
+    "use strict";
+
+
+    /**
+     * Break an expression into tokens defined in Twig.expression.definitions.
+     *
+     * @param {string} expression The string to tokenize.
+     *
+     * @return {Array} An array of tokens.
+     */
+    Twig.expression.tokenize = function (expression) {
+        var tokens = [],
+            // Keep an offset of the location in the expression for error messages.
+            exp_offset = 0,
+            // The valid next tokens of the previous token
+            next = null,
+            // Match information
+            type, regex, regex_i,
+            // The possible next token for the match
+            token_next,
+            // Has a match been found from the definitions
+            match_found, invalid_matches = [], match_function;
+
+        match_function = function () {
+            // Don't pass arguments to `Array.slice`, that is a performance killer
+            var match_i = arguments.length - 2, match = new Array(match_i);
+            while (match_i-- > 0) match[match_i] = arguments[match_i];
+
+            Twig.log.trace("Twig.expression.tokenize",
+                "Matched a ", type, " regular expression of ", match);
+
+            if (next && Twig.indexOf(next, type) < 0) {
+                invalid_matches.push(
+                    type + " cannot follow a " + tokens[tokens.length - 1].type +
+                    " at template:" + exp_offset + " near '" + match[0].substring(0, 20) +
+                    "...'"
+                );
+
+                // Not a match, don't change the expression
+                return match[0];
+            }
+
+            var handler = Twig.expression.handler[type];
+
+            // Validate the token if a validation function is provided
+            if (handler.validate && !handler.validate(match, tokens)) {
+                return match[0];
+            }
+
+            invalid_matches = [];
+
+            tokens.push({
+                type:  type,
+                value: match[0],
+                match: match
+            });
+
+            match_found = true;
+            next = token_next;
+            exp_offset += match[0].length;
+
+            // Does the token need to return output back to the expression string
+            // e.g. a function match of cycle( might return the '(' back to the expression
+            // This allows look-ahead to differentiate between token types (e.g. functions and variable names)
+            if (handler.transform) {
+                return handler.transform(match, tokens);
+            }
+            return '';
+        };
+
+        Twig.log.debug("Twig.expression.tokenize", "Tokenizing expression ", expression);
+
+        while (expression.length > 0) {
+            expression = expression.trim();
+            for (type in Twig.expression.handler) {
+                token_next = Twig.expression.handler[type].next;
+                regex = Twig.expression.handler[type].regex;
+                Twig.log.trace("Checking type ", type, " on ", expression);
+
+                match_found = false;
+
+                if (Twig.lib.isArray(regex)) {
+                    regex_i = regex.length;
+                    while (regex_i-- > 0)
+                        expression = expression.replace(regex[regex_i], match_function);
+                } else {
+                    expression = expression.replace(regex, match_function);
+                }
+
+                // An expression token has been matched. Break the for loop and start trying to
+                //  match the next template (if expression isn't empty.)
+                if (match_found) {
+                    break;
+                }
+            }
+            if (!match_found) {
+                if (invalid_matches.length > 0) {
+                    throw new Twig.Error(invalid_matches.join(" OR "));
+                } else {
+                    throw new Twig.Error("Unable to parse '" + expression + "' at template position" + exp_offset);
+                }
+            }
+        }
+
+        Twig.log.trace("Twig.expression.tokenize", "Tokenized to ", tokens);
+        return tokens;
+    };
+
+    /**
+     * Compile an expression token.
+     *
+     * @param {Object} raw_token The uncompiled token.
+     *
+     * @return {Object} The compiled token.
+     */
+    Twig.expression.compile = function (raw_token) {
+        var expression = raw_token.value,
+            // Tokenize expression
+            tokens = Twig.expression.tokenize(expression),
+            token = null,
+            output = [],
+            stack = [],
+            token_template = null;
+
+        Twig.log.trace("Twig.expression.compile: ", "Compiling ", expression);
+
+        // Push tokens into RPN stack using the Shunting-yard algorithm
+        // See http://en.wikipedia.org/wiki/Shunting_yard_algorithm
+
+        while (tokens.length > 0) {
+            token = tokens.shift();
+            token_template = Twig.expression.handler[token.type];
+
+            Twig.log.trace("Twig.expression.compile: ", "Compiling ", token);
+
+            // Compile the template
+            token_template.compile && token_template.compile(token, stack, output);
+
+            Twig.log.trace("Twig.expression.compile: ", "Stack is", stack);
+            Twig.log.trace("Twig.expression.compile: ", "Output is", output);
+        }
+
+        while(stack.length > 0) {
+            output.push(stack.pop());
+        }
+
+        Twig.log.trace("Twig.expression.compile: ", "Final output is", output);
+
+        raw_token.stack = output;
+        delete raw_token.value;
+
+        return raw_token;
+    };
+
+    return Twig;
+
+};
