+++ src/twig.expression.definitions.compile.js
@@ -1,0 +1,599 @@
+// ## twig.expression.definitions.compile.js
+//
+// This file handles tokenizing and compiling expressions.
+module.exports = function (Twig) {
+    "use strict";
+
+    // The regular expressions and compile logic used to match tokens in expressions.
+    //
+    // Properties:
+    //
+    //     type: The type of expression this matches
+    //
+    //     regex: One or more regular expressions that matche the format of the token.
+    //
+    //     next: Valid tokens that can occur next in the expression.
+    //
+    // Functions:
+    //
+    //     compile: A function that compiles the raw regular expression match into a token.
+    //
+    var definitions = [
+        {
+            type: Twig.expression.type.test,
+            regex: /^is\s+(not)?\s*([a-zA-Z_][a-zA-Z0-9_]*(\s?as)?)/,
+            next: Twig.expression.set.operations.concat([Twig.expression.type.parameter.start]),
+            compile: function(token, stack, output) {
+                token.filter     = token.match[2];
+                token.modifier = token.match[1];
+                delete token.match;
+                delete token.value;
+                output.push(token);
+            }
+        },
+        {
+            type: Twig.expression.type.comma,
+            // Match a comma
+            regex: /^,/,
+            next: Twig.expression.set.expressions.concat([Twig.expression.type.array.end, Twig.expression.type.object.end]),
+            compile: function(token, stack, output) {
+                var i = stack.length - 1,
+                    stack_token;
+
+                delete token.match;
+                delete token.value;
+
+                // pop tokens off the stack until the start of the object
+                for(;i >= 0; i--) {
+                    stack_token = stack.pop();
+                    if (stack_token.type === Twig.expression.type.object.start
+                        || stack_token.type === Twig.expression.type.parameter.start
+                        || stack_token.type === Twig.expression.type.array.start) {
+                        stack.push(stack_token);
+                        break;
+                    }
+                    output.push(stack_token);
+                }
+                output.push(token);
+            }
+        },
+        {
+            /**
+             * Match a number (integer or decimal)
+             */
+            type: Twig.expression.type.number,
+            // match a number
+            regex: /^\-?\d+(\.\d+)?/,
+            next: Twig.expression.set.operations,
+            compile: function(token, stack, output) {
+                token.value = Number(token.value);
+                output.push(token);
+            }
+        },
+        {
+            type: Twig.expression.type.operator.binary,
+            // Match any of ?:, +, *, /, -, %, ~, <, <=, >, >=, !=, ==, **, ?, :, and, b-and, or, b-or, b-xor, in, not in
+            // and, or, in, not in can be followed by a space or parenthesis
+            regex: /(^\?\:|^(b\-and)|^(b\-or)|^(b\-xor)|^[\+\-~%\?]|^[\:](?!\d\])|^[!=]==?|^[!<>]=?|^\*\*?|^\/\/?|^(and)[\(|\s+]|^(or)[\(|\s+]|^(in)[\(|\s+]|^(not in)[\(|\s+]|^\.\.)/,
+            next: Twig.expression.set.expressions,
+            transform: function(match, tokens) {
+                switch(match[0]) {
+                    case 'and(':
+                    case 'or(':
+                    case 'in(':
+                    case 'not in(':
+                        //Strip off the ( if it exists
+                        tokens[tokens.length - 1].value = match[2];
+                        return match[0];
+                        break;
+                    default:
+                        return '';
+                }
+            },
+            compile: function(token, stack, output) {
+                delete token.match;
+
+                token.value = token.value.trim();
+                var value = token.value,
+                    operator = Twig.expression.operator.lookup(value, token);
+
+                Twig.log.trace("Twig.expression.compile: ", "Operator: ", operator, " from ", value);
+
+                while (stack.length > 0 &&
+                (stack[stack.length-1].type == Twig.expression.type.operator.unary || stack[stack.length-1].type == Twig.expression.type.operator.binary) &&
+                (
+                    (operator.associativity === Twig.expression.operator.leftToRight &&
+                    operator.precidence        >= stack[stack.length-1].precidence) ||
+
+                    (operator.associativity === Twig.expression.operator.rightToLeft &&
+                    operator.precidence        >    stack[stack.length-1].precidence)
+                )
+                    ) {
+                    var temp = stack.pop();
+                    output.push(temp);
+                }
+
+                if (value === ":") {
+                    // Check if this is a ternary or object key being set
+                    if (stack[stack.length - 1] && stack[stack.length-1].value === "?") {
+                        // Continue as normal for a ternary
+                    } else {
+                        // This is not a ternary so we push the token to the output where it can be handled
+                        //     when the assocated object is closed.
+                        var key_token = output.pop();
+
+                        if (key_token.type === Twig.expression.type.string ||
+                            key_token.type === Twig.expression.type.variable) {
+                            token.key = key_token.value;
+                        } else if (key_token.type === Twig.expression.type.number) {
+                            // Convert integer keys into string keys
+                            token.key = key_token.value.toString();
+                        } else if (key_token.expression &&
+                            (key_token.type === Twig.expression.type.parameter.end ||
+                            key_token.type == Twig.expression.type.subexpression.end)) {
+                            token.params = key_token.params;
+                        } else {
+                            throw new Twig.Error("Unexpected value before ':' of " + key_token.type + " = " + key_token.value);
+                        }
+
+                        output.push(token);
+                        return;
+                    }
+                } else {
+                    stack.push(operator);
+                }
+            }
+        },
+        {
+            type: Twig.expression.type.operator.unary,
+            // Match any of not
+            regex: /(^not\s+)/,
+            next: Twig.expression.set.expressions,
+            compile: function(token, stack, output) {
+                delete token.match;
+
+                token.value = token.value.trim();
+                var value = token.value,
+                    operator = Twig.expression.operator.lookup(value, token);
+
+                Twig.log.trace("Twig.expression.compile: ", "Operator: ", operator, " from ", value);
+
+                while (stack.length > 0 &&
+                (stack[stack.length-1].type == Twig.expression.type.operator.unary || stack[stack.length-1].type == Twig.expression.type.operator.binary) &&
+                (
+                    (operator.associativity === Twig.expression.operator.leftToRight &&
+                    operator.precidence        >= stack[stack.length-1].precidence) ||
+
+                    (operator.associativity === Twig.expression.operator.rightToLeft &&
+                    operator.precidence        >    stack[stack.length-1].precidence)
+                )
+                    ) {
+                    var temp = stack.pop();
+                    output.push(temp);
+                }
+
+                stack.push(operator);
+            }
+        },
+        {
+            /**
+             * Match a string. This is anything between a pair of single or double quotes.
+             */
+            type: Twig.expression.type.string,
+            // See: http://blog.stevenlevithan.com/archives/match-quoted-string
+            regex: /^(["'])(?:(?=(\\?))\2[\s\S])*?\1/,
+            next: Twig.expression.set.operations_extended,
+            compile: function(token, stack, output) {
+                var value = token.value;
+                delete token.match
+
+                // Remove the quotes from the string
+                if (value.substring(0, 1) === '"') {
+                    value = value.replace('\\"', '"');
+                } else {
+                    value = value.replace("\\'", "'");
+                }
+                token.value = value.substring(1, value.length-1).replace( /\\n/g, "\n" ).replace( /\\r/g, "\r" );
+                Twig.log.trace("Twig.expression.compile: ", "String value: ", token.value);
+                output.push(token);
+            }
+        },
+        {
+            /**
+             * Match a subexpression set start.
+             */
+            type: Twig.expression.type.subexpression.start,
+            regex: /^\(/,
+            next: Twig.expression.set.expressions.concat([Twig.expression.type.subexpression.end]),
+            compile: function(token, stack, output) {
+                token.value = '(';
+                output.push(token);
+                stack.push(token);
+            }
+        },
+        {
+            /**
+             * Match a subexpression set end.
+             */
+            type: Twig.expression.type.subexpression.end,
+            regex: /^\)/,
+            next: Twig.expression.set.operations_extended,
+            validate: function(match, tokens) {
+                // Iterate back through previous tokens to ensure we follow a subexpression start
+                var i = tokens.length - 1,
+                    found_subexpression_start = false,
+                    next_subexpression_start_invalid = false,
+                    unclosed_parameter_count = 0;
+
+                while(!found_subexpression_start && i >= 0) {
+                    var token = tokens[i];
+
+                    found_subexpression_start = token.type === Twig.expression.type.subexpression.start;
+
+                    // If we have previously found a subexpression end, then this subexpression start is the start of
+                    // that subexpression, not the subexpression we are searching for
+                    if (found_subexpression_start && next_subexpression_start_invalid) {
+                        next_subexpression_start_invalid = false;
+                        found_subexpression_start = false;
+                    }
+
+                    // Count parameter tokens to ensure we dont return truthy for a parameter opener
+                    if (token.type === Twig.expression.type.parameter.start) {
+                        unclosed_parameter_count++;
+                    } else if (token.type === Twig.expression.type.parameter.end) {
+                        unclosed_parameter_count--;
+                    } else if (token.type === Twig.expression.type.subexpression.end) {
+                        next_subexpression_start_invalid = true;
+                    }
+
+                    i--;
+                }
+
+                // If we found unclosed parameters, return false
+                // If we didnt find subexpression start, return false
+                // Otherwise return true
+
+                return (found_subexpression_start && (unclosed_parameter_count === 0));
+            },
+            compile: function(token, stack, output) {
+                // This is basically a copy of parameter end compilation
+                var stack_token,
+                    end_token = token;
+
+                stack_token = stack.pop();
+                while(stack.length > 0 && stack_token.type != Twig.expression.type.subexpression.start) {
+                    output.push(stack_token);
+                    stack_token = stack.pop();
+                }
+
+                // Move contents of parens into preceding filter
+                var param_stack = [];
+                while(token.type !== Twig.expression.type.subexpression.start) {
+                    // Add token to arguments stack
+                    param_stack.unshift(token);
+                    token = output.pop();
+                }
+
+                param_stack.unshift(token);
+
+                var is_expression = false;
+
+                //If the token at the top of the *stack* is a function token, pop it onto the output queue.
+                // Get the token preceding the parameters
+                stack_token = stack[stack.length-1];
+
+                if (stack_token === undefined ||
+                    (stack_token.type !== Twig.expression.type._function &&
+                    stack_token.type !== Twig.expression.type.filter &&
+                    stack_token.type !== Twig.expression.type.test &&
+                    stack_token.type !== Twig.expression.type.key.brackets)) {
+
+                    end_token.expression = true;
+
+                    // remove start and end token from stack
+                    param_stack.pop();
+                    param_stack.shift();
+
+                    end_token.params = param_stack;
+
+                    output.push(end_token);
+                } else {
+                    // This should never be hit
+                    end_token.expression = false;
+                    stack_token.params = param_stack;
+                }
+            }
+        },
+        {
+            /**
+             * Match a parameter set start.
+             */
+            type: Twig.expression.type.parameter.start,
+            regex: /^\(/,
+            next: Twig.expression.set.expressions.concat([Twig.expression.type.parameter.end]),
+            validate: function(match, tokens) {
+                var last_token = tokens[tokens.length - 1];
+                // We can't use the regex to test if we follow a space because expression is trimmed
+                return last_token && (Twig.indexOf(Twig.expression.reservedWords, last_token.value.trim()) < 0);
+            },
+            compile: Twig.expression.fn.compile.push_both
+        },
+        {
+            /**
+             * Match a parameter set end.
+             */
+            type: Twig.expression.type.parameter.end,
+            regex: /^\)/,
+            next: Twig.expression.set.operations_extended,
+            compile: function(token, stack, output) {
+                var stack_token,
+                    end_token = token;
+
+                stack_token = stack.pop();
+                while(stack.length > 0 && stack_token.type != Twig.expression.type.parameter.start) {
+                    output.push(stack_token);
+                    stack_token = stack.pop();
+                }
+
+                // Move contents of parens into preceding filter
+                var param_stack = [];
+                while(token.type !== Twig.expression.type.parameter.start) {
+                    // Add token to arguments stack
+                    param_stack.unshift(token);
+                    token = output.pop();
+                }
+                param_stack.unshift(token);
+
+                var is_expression = false;
+
+                // Get the token preceding the parameters
+                token = output[output.length-1];
+
+                if (token === undefined ||
+                    (token.type !== Twig.expression.type._function &&
+                    token.type !== Twig.expression.type.filter &&
+                    token.type !== Twig.expression.type.test &&
+                    token.type !== Twig.expression.type.key.brackets)) {
+
+                    end_token.expression = true;
+
+                    // remove start and end token from stack
+                    param_stack.pop();
+                    param_stack.shift();
+
+                    end_token.params = param_stack;
+
+                    output.push(end_token);
+
+                } else {
+                    end_token.expression = false;
+                    token.params = param_stack;
+                }
+            }
+        },
+        {
+            type: Twig.expression.type.slice,
+            regex: /^\[(\d*\:\d*)\]/,
+            next: Twig.expression.set.operations_extended,
+            compile: function(token, stack, output) {
+                var sliceRange = token.match[1].split(':');
+
+                //sliceStart can be undefined when we pass parameters to the slice filter later
+                var sliceStart = (sliceRange[0]) ? parseInt(sliceRange[0]) : undefined;
+                var sliceEnd = (sliceRange[1]) ? parseInt(sliceRange[1]) : undefined;
+
+                token.value = 'slice';
+                token.params = [sliceStart, sliceEnd];
+
+                //sliceEnd can't be undefined as the slice filter doesn't check for this, but it does check the length
+                //of the params array, so just shorten it.
+                if (!sliceEnd) {
+                    token.params = [sliceStart];
+                }
+
+                output.push(token);
+            }
+        },
+        {
+            /**
+             * Match an array start.
+             */
+            type: Twig.expression.type.array.start,
+            regex: /^\[/,
+            next: Twig.expression.set.expressions.concat([Twig.expression.type.array.end]),
+            compile: Twig.expression.fn.compile.push_both
+        },
+        {
+            /**
+             * Match an array end.
+             */
+            type: Twig.expression.type.array.end,
+            regex: /^\]/,
+            next: Twig.expression.set.operations_extended,
+            compile: function(token, stack, output) {
+                var i = stack.length - 1,
+                    stack_token;
+                // pop tokens off the stack until the start of the object
+                for(;i >= 0; i--) {
+                    stack_token = stack.pop();
+                    if (stack_token.type === Twig.expression.type.array.start) {
+                        break;
+                    }
+                    output.push(stack_token);
+                }
+                output.push(token);
+            }
+        },
+        // Token that represents the start of a hash map '}'
+        //
+        // Hash maps take the form:
+        //        { "key": 'value', "another_key": item }
+        //
+        // Keys must be quoted (either single or double) and values can be any expression.
+        {
+            type: Twig.expression.type.object.start,
+            regex: /^\{/,
+            next: Twig.expression.set.expressions.concat([Twig.expression.type.object.end]),
+            compile: Twig.expression.fn.compile.push_both
+        },
+
+        // Token that represents the end of a Hash Map '}'
+        //
+        // This is where the logic for building the internal
+        // representation of a hash map is defined.
+        {
+            type: Twig.expression.type.object.end,
+            regex: /^\}/,
+            next: Twig.expression.set.operations_extended,
+            compile: function(token, stack, output) {
+                var i = stack.length-1,
+                    stack_token;
+
+                // pop tokens off the stack until the start of the object
+                for(;i >= 0; i--) {
+                    stack_token = stack.pop();
+                    if (stack_token && stack_token.type === Twig.expression.type.object.start) {
+                        break;
+                    }
+                    output.push(stack_token);
+                }
+                output.push(token);
+            }
+        },
+
+        // Token representing a filter
+        //
+        // Filters can follow any expression and take the form:
+        //        expression|filter(optional, args)
+        //
+        // Filter parsing is done in the Twig.filters namespace.
+        {
+            type: Twig.expression.type.filter,
+            // match a | then a letter or _, then any number of letters, numbers, _ or -
+            regex: /^\|\s?([a-zA-Z_][a-zA-Z0-9_\-]*)/,
+            next: Twig.expression.set.operations_extended.concat([
+                Twig.expression.type.parameter.start]),
+            compile: function(token, stack, output) {
+                token.value = token.match[1];
+                output.push(token);
+            }
+        },
+        {
+            type: Twig.expression.type._function,
+            // match any letter or _, then any number of letters, numbers, _ or - followed by (
+            regex: /^([a-zA-Z_][a-zA-Z0-9_]*)\s*\(/,
+            next: Twig.expression.type.parameter.start,
+            validate: function(match, tokens) {
+                // Make sure this function is not a reserved word
+                return match[1] && (Twig.indexOf(Twig.expression.reservedWords, match[1]) < 0);
+            },
+            transform: function(match, tokens) {
+                return '(';
+            },
+            compile: function(token, stack, output) {
+                var fn = token.match[1];
+                token.fn = fn;
+                // cleanup token
+                delete token.match;
+                delete token.value;
+
+                output.push(token);
+            }
+        },
+
+        // Token representing a variable.
+        //
+        // Variables can contain letters, numbers, underscores and
+        // dashes, but must start with a letter or underscore.
+        //
+        // Variables are retrieved from the render context and take
+        // the value of 'undefined' if the given variable doesn't
+        // exist in the context.
+        {
+            type: Twig.expression.type.variable,
+            // match any letter or _, then any number of letters, numbers, _ or -
+            regex: /^[a-zA-Z_][a-zA-Z0-9_]*/,
+            next: Twig.expression.set.operations_extended.concat([
+                Twig.expression.type.parameter.start]),
+            compile: Twig.expression.fn.compile.push,
+            validate: function(match, tokens) {
+                return (Twig.indexOf(Twig.expression.reservedWords, match[0]) < 0);
+            }
+        },
+        {
+            type: Twig.expression.type.key.period,
+            regex: /^\.([a-zA-Z0-9_]+)/,
+            next: Twig.expression.set.operations_extended.concat([
+                Twig.expression.type.parameter.start]),
+            compile: function(token, stack, output) {
+                token.key = token.match[1];
+                delete token.match;
+                delete token.value;
+
+                output.push(token);
+            }
+        },
+        {
+            type: Twig.expression.type.key.brackets,
+            regex: /^\[([^\]\:]*)\]/,
+            next: Twig.expression.set.operations_extended.concat([
+                Twig.expression.type.parameter.start]),
+            compile: function(token, stack, output) {
+                var match = token.match[1];
+                delete token.value;
+                delete token.match;
+
+                // The expression stack for the key
+                token.stack = Twig.expression.compile({
+                    value: match
+                }).stack;
+
+                output.push(token);
+            }
+        },
+        {
+            /**
+             * Match a null value.
+             */
+            type: Twig.expression.type._null,
+            // match a number
+            regex: /^(null|NULL|none|NONE)/,
+            next: Twig.expression.set.operations,
+            compile: function(token, stack, output) {
+                delete token.match;
+                token.value = null;
+                output.push(token);
+            }
+        },
+        {
+            /**
+             * Match the context
+             */
+            type: Twig.expression.type.context,
+            regex: /^_context/,
+            next: Twig.expression.set.operations_extended.concat([
+                Twig.expression.type.parameter.start]),
+            compile: Twig.expression.fn.compile.push
+        },
+        {
+            /**
+             * Match a boolean
+             */
+            type: Twig.expression.type.bool,
+            regex: /^(true|TRUE|false|FALSE)/,
+            next: Twig.expression.set.operations,
+            compile: function(token, stack, output) {
+                token.value = (token.match[0].toLowerCase( ) === "true");
+                delete token.match;
+                output.push(token);
+            }
+        }
+    ];
+
+    definitions.forEach(function (definition) {
+        Twig.expression.extend(definition);
+    });
+
+    return Twig;
+
+};
