+++ src/twig.expression.js
@@ -1,22 +1,13 @@
 // ## twig.expression.js
 //
-// This file handles tokenizing, compiling and parsing expressions.
+// This file handles tokenizing and parsing expressions.
 module.exports = function (Twig) {
     "use strict";
-
-    function parseParams(thisArg, params, context) {
-        if (params)
-            return Twig.expression.parseAsync.call(thisArg, params, context);
-
-        return Twig.Promise.resolve(false);
-    }
 
     /**
      * Namespace for expression handling.
      */
     Twig.expression = { };
-
-    require('./twig.expression.operator')(Twig);
 
     /**
      * Reserved word that can't be used as variable names.
@@ -122,873 +113,6 @@ module.exports = function (Twig) {
         }
     };
 
-    // The regular expressions and compile/parse logic used to match tokens in expressions.
-    //
-    // Properties:
-    //
-    //      type:  The type of expression this matches
-    //
-    //      regex: One or more regular expressions that matche the format of the token.
-    //
-    //      next:  Valid tokens that can occur next in the expression.
-    //
-    // Functions:
-    //
-    //      compile: A function that compiles the raw regular expression match into a token.
-    //
-    //      parse:   A function that parses the compiled token into output.
-    //
-    Twig.expression.definitions = [
-        {
-            type: Twig.expression.type.test,
-            regex: /^is\s+(not)?\s*([a-zA-Z_][a-zA-Z0-9_]*(\s?as)?)/,
-            next: Twig.expression.set.operations.concat([Twig.expression.type.parameter.start]),
-            compile: function(token, stack, output) {
-                token.filter   = token.match[2];
-                token.modifier = token.match[1];
-                delete token.match;
-                delete token.value;
-                output.push(token);
-            },
-            parse: function(token, stack, context) {
-                var value = stack.pop();
-
-                return parseParams(this, token.params, context)
-                .then(function(params) {
-                    var result = Twig.test(token.filter, value, params);
-
-                    if (token.modifier == 'not') {
-                        stack.push(!result);
-                    } else {
-                        stack.push(result);
-                    }
-                });
-            }
-        },
-        {
-            type: Twig.expression.type.comma,
-            // Match a comma
-            regex: /^,/,
-            next: Twig.expression.set.expressions.concat([Twig.expression.type.array.end, Twig.expression.type.object.end]),
-            compile: function(token, stack, output) {
-                var i = stack.length - 1,
-                    stack_token;
-
-                delete token.match;
-                delete token.value;
-
-                // pop tokens off the stack until the start of the object
-                for(;i >= 0; i--) {
-                    stack_token = stack.pop();
-                    if (stack_token.type === Twig.expression.type.object.start
-                            || stack_token.type === Twig.expression.type.parameter.start
-                            || stack_token.type === Twig.expression.type.array.start) {
-                        stack.push(stack_token);
-                        break;
-                    }
-                    output.push(stack_token);
-                }
-                output.push(token);
-            }
-        },
-        {
-            /**
-             * Match a number (integer or decimal)
-             */
-            type: Twig.expression.type.number,
-            // match a number
-            regex: /^\-?\d+(\.\d+)?/,
-            next: Twig.expression.set.operations,
-            compile: function(token, stack, output) {
-                token.value = Number(token.value);
-                output.push(token);
-            },
-            parse: Twig.expression.fn.parse.push_value
-        },
-        {
-            type: Twig.expression.type.operator.binary,
-            // Match any of ?:, +, *, /, -, %, ~, <, <=, >, >=, !=, ==, **, ?, :, and, b-and, or, b-or, b-xor, in, not in
-            // and, or, in, not in can be followed by a space or parenthesis
-            regex: /(^\?\:|^(b\-and)|^(b\-or)|^(b\-xor)|^[\+\-~%\?]|^[\:](?!\d\])|^[!=]==?|^[!<>]=?|^\*\*?|^\/\/?|^(and)[\(|\s+]|^(or)[\(|\s+]|^(in)[\(|\s+]|^(not in)[\(|\s+]|^\.\.)/,
-            next: Twig.expression.set.expressions,
-            transform: function(match, tokens) {
-                switch(match[0]) {
-                    case 'and(':
-                    case 'or(':
-                    case 'in(':
-                    case 'not in(':
-                        //Strip off the ( if it exists
-                        tokens[tokens.length - 1].value = match[2];
-                        return match[0];
-                        break;
-                    default:
-                        return '';
-                }
-            },
-            compile: function(token, stack, output) {
-                delete token.match;
-
-                token.value = token.value.trim();
-                var value = token.value,
-                    operator = Twig.expression.operator.lookup(value, token);
-
-                Twig.log.trace("Twig.expression.compile: ", "Operator: ", operator, " from ", value);
-
-                while (stack.length > 0 &&
-                       (stack[stack.length-1].type == Twig.expression.type.operator.unary || stack[stack.length-1].type == Twig.expression.type.operator.binary) &&
-                            (
-                                (operator.associativity === Twig.expression.operator.leftToRight &&
-                                 operator.precidence    >= stack[stack.length-1].precidence) ||
-
-                                (operator.associativity === Twig.expression.operator.rightToLeft &&
-                                 operator.precidence    >  stack[stack.length-1].precidence)
-                            )
-                       ) {
-                     var temp = stack.pop();
-                     output.push(temp);
-                }
-
-                if (value === ":") {
-                    // Check if this is a ternary or object key being set
-                    if (stack[stack.length - 1] && stack[stack.length-1].value === "?") {
-                        // Continue as normal for a ternary
-                    } else {
-                        // This is not a ternary so we push the token to the output where it can be handled
-                        //   when the assocated object is closed.
-                        var key_token = output.pop();
-
-                        if (key_token.type === Twig.expression.type.string ||
-                                key_token.type === Twig.expression.type.variable) {
-                            token.key = key_token.value;
-                        } else if (key_token.type === Twig.expression.type.number) {
-                            // Convert integer keys into string keys
-                            token.key = key_token.value.toString();
-                        } else if (key_token.expression &&
-                            (key_token.type === Twig.expression.type.parameter.end ||
-                            key_token.type == Twig.expression.type.subexpression.end)) {
-                            token.params = key_token.params;
-                        } else {
-                            throw new Twig.Error("Unexpected value before ':' of " + key_token.type + " = " + key_token.value);
-                        }
-
-                        output.push(token);
-                        return;
-                    }
-                } else {
-                    stack.push(operator);
-                }
-            },
-            parse: function(token, stack, context) {
-                if (token.key) {
-                    // handle ternary ':' operator
-                    stack.push(token);
-                } else if (token.params) {
-                    // handle "{(expression):value}"
-                    return Twig.expression.parseAsync.call(this, token.params, context)
-                    .then(function(key) {
-                        token.key = key;
-                        stack.push(token);
-
-                        //If we're in a loop, we might need token.params later, especially in this form of "(expression):value"
-                        if (!context.loop) {
-                            delete(token.params);
-                        }
-                    });
-                } else {
-                    Twig.expression.operator.parse(token.value, stack);
-                }
-            }
-        },
-        {
-            type: Twig.expression.type.operator.unary,
-            // Match any of not
-            regex: /(^not\s+)/,
-            next: Twig.expression.set.expressions,
-            compile: function(token, stack, output) {
-                delete token.match;
-
-                token.value = token.value.trim();
-                var value = token.value,
-                    operator = Twig.expression.operator.lookup(value, token);
-
-                Twig.log.trace("Twig.expression.compile: ", "Operator: ", operator, " from ", value);
-
-                while (stack.length > 0 &&
-                       (stack[stack.length-1].type == Twig.expression.type.operator.unary || stack[stack.length-1].type == Twig.expression.type.operator.binary) &&
-                            (
-                                (operator.associativity === Twig.expression.operator.leftToRight &&
-                                 operator.precidence    >= stack[stack.length-1].precidence) ||
-
-                                (operator.associativity === Twig.expression.operator.rightToLeft &&
-                                 operator.precidence    >  stack[stack.length-1].precidence)
-                            )
-                       ) {
-                     var temp = stack.pop();
-                     output.push(temp);
-                }
-
-                stack.push(operator);
-            },
-            parse: function(token, stack, context) {
-                Twig.expression.operator.parse(token.value, stack);
-            }
-        },
-        {
-            /**
-             * Match a string. This is anything between a pair of single or double quotes.
-             */
-            type: Twig.expression.type.string,
-            // See: http://blog.stevenlevithan.com/archives/match-quoted-string
-            regex: /^(["'])(?:(?=(\\?))\2[\s\S])*?\1/,
-            next: Twig.expression.set.operations_extended,
-            compile: function(token, stack, output) {
-                var value = token.value;
-                delete token.match
-
-                // Remove the quotes from the string
-                if (value.substring(0, 1) === '"') {
-                    value = value.replace('\\"', '"');
-                } else {
-                    value = value.replace("\\'", "'");
-                }
-                token.value = value.substring(1, value.length-1).replace( /\\n/g, "\n" ).replace( /\\r/g, "\r" );
-                Twig.log.trace("Twig.expression.compile: ", "String value: ", token.value);
-                output.push(token);
-            },
-            parse: Twig.expression.fn.parse.push_value
-        },
-        {
-            /**
-             * Match a subexpression set start.
-             */
-            type: Twig.expression.type.subexpression.start,
-            regex: /^\(/,
-            next: Twig.expression.set.expressions.concat([Twig.expression.type.subexpression.end]),
-            compile: function(token, stack, output) {
-                token.value = '(';
-                output.push(token);
-                stack.push(token);
-            },
-            parse: Twig.expression.fn.parse.push
-        },
-        {
-            /**
-             * Match a subexpression set end.
-             */
-            type: Twig.expression.type.subexpression.end,
-            regex: /^\)/,
-            next: Twig.expression.set.operations_extended,
-            validate: function(match, tokens) {
-                // Iterate back through previous tokens to ensure we follow a subexpression start
-                var i = tokens.length - 1,
-                    found_subexpression_start = false,
-                    next_subexpression_start_invalid = false,
-                    unclosed_parameter_count = 0;
-
-                while(!found_subexpression_start && i >= 0) {
-                    var token = tokens[i];
-
-                    found_subexpression_start = token.type === Twig.expression.type.subexpression.start;
-
-                    // If we have previously found a subexpression end, then this subexpression start is the start of
-                    // that subexpression, not the subexpression we are searching for
-                    if (found_subexpression_start && next_subexpression_start_invalid) {
-                        next_subexpression_start_invalid = false;
-                        found_subexpression_start = false;
-                    }
-
-                    // Count parameter tokens to ensure we dont return truthy for a parameter opener
-                    if (token.type === Twig.expression.type.parameter.start) {
-                        unclosed_parameter_count++;
-                    } else if (token.type === Twig.expression.type.parameter.end) {
-                        unclosed_parameter_count--;
-                    } else if (token.type === Twig.expression.type.subexpression.end) {
-                        next_subexpression_start_invalid = true;
-                    }
-
-                    i--;
-                }
-
-                // If we found unclosed parameters, return false
-                // If we didnt find subexpression start, return false
-                // Otherwise return true
-
-                return (found_subexpression_start && (unclosed_parameter_count === 0));
-            },
-            compile: function(token, stack, output) {
-                // This is basically a copy of parameter end compilation
-                var stack_token,
-                    end_token = token;
-
-                stack_token = stack.pop();
-                while(stack.length > 0 && stack_token.type != Twig.expression.type.subexpression.start) {
-                    output.push(stack_token);
-                    stack_token = stack.pop();
-                }
-
-                // Move contents of parens into preceding filter
-                var param_stack = [];
-                while(token.type !== Twig.expression.type.subexpression.start) {
-                    // Add token to arguments stack
-                    param_stack.unshift(token);
-                    token = output.pop();
-                }
-
-                param_stack.unshift(token);
-
-                var is_expression = false;
-
-                //If the token at the top of the *stack* is a function token, pop it onto the output queue.
-                // Get the token preceding the parameters
-                stack_token = stack[stack.length-1];
-
-                if (stack_token === undefined ||
-                    (stack_token.type !== Twig.expression.type._function &&
-                    stack_token.type !== Twig.expression.type.filter &&
-                    stack_token.type !== Twig.expression.type.test &&
-                    stack_token.type !== Twig.expression.type.key.brackets)) {
-
-                    end_token.expression = true;
-
-                    // remove start and end token from stack
-                    param_stack.pop();
-                    param_stack.shift();
-
-                    end_token.params = param_stack;
-
-                    output.push(end_token);
-                } else {
-                    // This should never be hit
-                    end_token.expression = false;
-                    stack_token.params = param_stack;
-                }
-            },
-            parse: function(token, stack, context) {
-                var new_array = [],
-                    array_ended = false,
-                    value = null;
-
-                if (token.expression) {
-                    return Twig.expression.parseAsync.call(this, token.params, context)
-                    .then(function(value) {
-                        stack.push(value);
-                    });
-                } else {
-                    throw new Twig.Error("Unexpected subexpression end when token is not marked as an expression");
-                }
-            }
-        },
-        {
-            /**
-             * Match a parameter set start.
-             */
-            type: Twig.expression.type.parameter.start,
-            regex: /^\(/,
-            next: Twig.expression.set.expressions.concat([Twig.expression.type.parameter.end]),
-            validate: function(match, tokens) {
-                var last_token = tokens[tokens.length - 1];
-                // We can't use the regex to test if we follow a space because expression is trimmed
-                return last_token && (Twig.indexOf(Twig.expression.reservedWords, last_token.value.trim()) < 0);
-            },
-            compile: Twig.expression.fn.compile.push_both,
-            parse: Twig.expression.fn.parse.push
-        },
-        {
-            /**
-             * Match a parameter set end.
-             */
-            type: Twig.expression.type.parameter.end,
-            regex: /^\)/,
-            next: Twig.expression.set.operations_extended,
-            compile: function(token, stack, output) {
-                var stack_token,
-                    end_token = token;
-
-                stack_token = stack.pop();
-                while(stack.length > 0 && stack_token.type != Twig.expression.type.parameter.start) {
-                    output.push(stack_token);
-                    stack_token = stack.pop();
-                }
-
-                // Move contents of parens into preceding filter
-                var param_stack = [];
-                while(token.type !== Twig.expression.type.parameter.start) {
-                    // Add token to arguments stack
-                    param_stack.unshift(token);
-                    token = output.pop();
-                }
-                param_stack.unshift(token);
-
-                var is_expression = false;
-
-                // Get the token preceding the parameters
-                token = output[output.length-1];
-
-                if (token === undefined ||
-                    (token.type !== Twig.expression.type._function &&
-                    token.type !== Twig.expression.type.filter &&
-                    token.type !== Twig.expression.type.test &&
-                    token.type !== Twig.expression.type.key.brackets)) {
-
-                    end_token.expression = true;
-
-                    // remove start and end token from stack
-                    param_stack.pop();
-                    param_stack.shift();
-
-                    end_token.params = param_stack;
-
-                    output.push(end_token);
-
-                } else {
-                    end_token.expression = false;
-                    token.params = param_stack;
-                }
-            },
-            parse: function(token, stack, context) {
-                var new_array = [],
-                    array_ended = false,
-                    value = null;
-
-                if (token.expression) {
-                    return Twig.expression.parseAsync.call(this, token.params, context)
-                    .then(function(value) {
-                        stack.push(value);
-                    });
-                } else {
-
-                    while (stack.length > 0) {
-                        value = stack.pop();
-                        // Push values into the array until the start of the array
-                        if (value && value.type && value.type == Twig.expression.type.parameter.start) {
-                            array_ended = true;
-                            break;
-                        }
-                        new_array.unshift(value);
-                    }
-
-                    if (!array_ended) {
-                        throw new Twig.Error("Expected end of parameter set.");
-                    }
-
-                    stack.push(new_array);
-                }
-            }
-        },
-        {
-            type: Twig.expression.type.slice,
-            regex: /^\[(\d*\:\d*)\]/,
-            next: Twig.expression.set.operations_extended,
-            compile: function(token, stack, output) {
-                var sliceRange = token.match[1].split(':');
-
-                //sliceStart can be undefined when we pass parameters to the slice filter later
-                var sliceStart = (sliceRange[0]) ? parseInt(sliceRange[0]) : undefined;
-                var sliceEnd = (sliceRange[1]) ? parseInt(sliceRange[1]) : undefined;
-
-                token.value = 'slice';
-                token.params = [sliceStart, sliceEnd];
-
-                //sliceEnd can't be undefined as the slice filter doesn't check for this, but it does check the length
-                //of the params array, so just shorten it.
-                if (!sliceEnd) {
-                    token.params = [sliceStart];
-                }
-
-                output.push(token);
-            },
-            parse: function(token, stack, context) {
-                var input = stack.pop(),
-                    params = token.params;
-
-                stack.push(Twig.filter.call(this, token.value, input, params));
-            }
-        },
-        {
-            /**
-             * Match an array start.
-             */
-            type: Twig.expression.type.array.start,
-            regex: /^\[/,
-            next: Twig.expression.set.expressions.concat([Twig.expression.type.array.end]),
-            compile: Twig.expression.fn.compile.push_both,
-            parse: Twig.expression.fn.parse.push
-        },
-        {
-            /**
-             * Match an array end.
-             */
-            type: Twig.expression.type.array.end,
-            regex: /^\]/,
-            next: Twig.expression.set.operations_extended,
-            compile: function(token, stack, output) {
-                var i = stack.length - 1,
-                    stack_token;
-                // pop tokens off the stack until the start of the object
-                for(;i >= 0; i--) {
-                    stack_token = stack.pop();
-                    if (stack_token.type === Twig.expression.type.array.start) {
-                        break;
-                    }
-                    output.push(stack_token);
-                }
-                output.push(token);
-            },
-            parse: function(token, stack, context) {
-                var new_array = [],
-                    array_ended = false,
-                    value = null;
-
-                while (stack.length > 0) {
-                    value = stack.pop();
-                    // Push values into the array until the start of the array
-                    if (value.type && value.type == Twig.expression.type.array.start) {
-                        array_ended = true;
-                        break;
-                    }
-                    new_array.unshift(value);
-                }
-                if (!array_ended) {
-                    throw new Twig.Error("Expected end of array.");
-                }
-
-                stack.push(new_array);
-            }
-        },
-        // Token that represents the start of a hash map '}'
-        //
-        // Hash maps take the form:
-        //    { "key": 'value', "another_key": item }
-        //
-        // Keys must be quoted (either single or double) and values can be any expression.
-        {
-            type: Twig.expression.type.object.start,
-            regex: /^\{/,
-            next: Twig.expression.set.expressions.concat([Twig.expression.type.object.end]),
-            compile: Twig.expression.fn.compile.push_both,
-            parse: Twig.expression.fn.parse.push
-        },
-
-        // Token that represents the end of a Hash Map '}'
-        //
-        // This is where the logic for building the internal
-        // representation of a hash map is defined.
-        {
-            type: Twig.expression.type.object.end,
-            regex: /^\}/,
-            next: Twig.expression.set.operations_extended,
-            compile: function(token, stack, output) {
-                var i = stack.length-1,
-                    stack_token;
-
-                // pop tokens off the stack until the start of the object
-                for(;i >= 0; i--) {
-                    stack_token = stack.pop();
-                    if (stack_token && stack_token.type === Twig.expression.type.object.start) {
-                        break;
-                    }
-                    output.push(stack_token);
-                }
-                output.push(token);
-            },
-            parse: function(end_token, stack, context) {
-                var new_object = {},
-                    object_ended = false,
-                    token = null,
-                    token_key = null,
-                    has_value = false,
-                    value = null;
-
-                while (stack.length > 0) {
-                    token = stack.pop();
-                    // Push values into the array until the start of the object
-                    if (token && token.type && token.type === Twig.expression.type.object.start) {
-                        object_ended = true;
-                        break;
-                    }
-                    if (token && token.type && (token.type === Twig.expression.type.operator.binary || token.type === Twig.expression.type.operator.unary) && token.key) {
-                        if (!has_value) {
-                            throw new Twig.Error("Missing value for key '" + token.key + "' in object definition.");
-                        }
-                        new_object[token.key] = value;
-
-                        // Preserve the order that elements are added to the map
-                        // This is necessary since JavaScript objects don't
-                        // guarantee the order of keys
-                        if (new_object._keys === undefined) new_object._keys = [];
-                        new_object._keys.unshift(token.key);
-
-                        // reset value check
-                        value = null;
-                        has_value = false;
-
-                    } else {
-                        has_value = true;
-                        value = token;
-                    }
-                }
-                if (!object_ended) {
-                    throw new Twig.Error("Unexpected end of object.");
-                }
-
-                stack.push(new_object);
-            }
-        },
-
-        // Token representing a filter
-        //
-        // Filters can follow any expression and take the form:
-        //    expression|filter(optional, args)
-        //
-        // Filter parsing is done in the Twig.filters namespace.
-        {
-            type: Twig.expression.type.filter,
-            // match a | then a letter or _, then any number of letters, numbers, _ or -
-            regex: /^\|\s?([a-zA-Z_][a-zA-Z0-9_\-]*)/,
-            next: Twig.expression.set.operations_extended.concat([
-                    Twig.expression.type.parameter.start]),
-            compile: function(token, stack, output) {
-                token.value = token.match[1];
-                output.push(token);
-            },
-            parse: function(token, stack, context) {
-                var that = this,
-                    input = stack.pop();
-
-                return parseParams(this, token.params, context)
-                .then(function(params) {
-                    return Twig.filter.call(that, token.value, input, params);
-                })
-                .then(function(value) {
-                    stack.push(value);
-                });
-            }
-        },
-        {
-            type: Twig.expression.type._function,
-            // match any letter or _, then any number of letters, numbers, _ or - followed by (
-            regex: /^([a-zA-Z_][a-zA-Z0-9_]*)\s*\(/,
-            next: Twig.expression.type.parameter.start,
-            validate: function(match, tokens) {
-                // Make sure this function is not a reserved word
-                return match[1] && (Twig.indexOf(Twig.expression.reservedWords, match[1]) < 0);
-            },
-            transform: function(match, tokens) {
-                return '(';
-            },
-            compile: function(token, stack, output) {
-                var fn = token.match[1];
-                token.fn = fn;
-                // cleanup token
-                delete token.match;
-                delete token.value;
-
-                output.push(token);
-            },
-            parse: function(token, stack, context) {
-
-                var that = this,
-                    fn = token.fn,
-                    value;
-
-                return parseParams(this, token.params, context)
-                .then(function(params) {
-                    if (Twig.functions[fn]) {
-                        // Get the function from the built-in functions
-                        value = Twig.functions[fn].apply(that, params);
-
-                    } else if (typeof context[fn] == 'function') {
-                        // Get the function from the user/context defined functions
-                        value = context[fn].apply(context, params);
-
-                    } else {
-                        throw new Twig.Error(fn + ' function does not exist and is not defined in the context');
-                    }
-
-                    return value;
-                })
-                .then(function(result) {
-                    stack.push(result);
-                });
-            }
-        },
-
-        // Token representing a variable.
-        //
-        // Variables can contain letters, numbers, underscores and
-        // dashes, but must start with a letter or underscore.
-        //
-        // Variables are retrieved from the render context and take
-        // the value of 'undefined' if the given variable doesn't
-        // exist in the context.
-        {
-            type: Twig.expression.type.variable,
-            // match any letter or _, then any number of letters, numbers, _ or -
-            regex: /^[a-zA-Z_][a-zA-Z0-9_]*/,
-            next: Twig.expression.set.operations_extended.concat([
-                    Twig.expression.type.parameter.start]),
-            compile: Twig.expression.fn.compile.push,
-            validate: function(match, tokens) {
-                return (Twig.indexOf(Twig.expression.reservedWords, match[0]) < 0);
-            },
-            parse: function(token, stack, context) {
-                // Get the variable from the context
-                return Twig.expression.resolveAsync.call(this, context[token.value], context)
-                .then(function(value) {
-                    stack.push(value);
-                });
-            }
-        },
-        {
-            type: Twig.expression.type.key.period,
-            regex: /^\.([a-zA-Z0-9_]+)/,
-            next: Twig.expression.set.operations_extended.concat([
-                    Twig.expression.type.parameter.start]),
-            compile: function(token, stack, output) {
-                token.key = token.match[1];
-                delete token.match;
-                delete token.value;
-
-                output.push(token);
-            },
-            parse: function(token, stack, context, next_token) {
-                var that = this,
-                    key = token.key,
-                    object = stack.pop(),
-                    value;
-
-                return parseParams(this, token.params, context)
-                .then(function(params) {
-                    if (object === null || object === undefined) {
-                        if (that.options.strict_variables) {
-                            throw new Twig.Error("Can't access a key " + key + " on an null or undefined object.");
-                        } else {
-                            value = undefined;
-                        }
-                    } else {
-                        var capitalize = function (value) {
-                            return value.substr(0, 1).toUpperCase() + value.substr(1);
-                        };
-
-                        // Get the variable from the context
-                        if (typeof object === 'object' && key in object) {
-                            value = object[key];
-                        } else if (object["get" + capitalize(key)] !== undefined) {
-                            value = object["get" + capitalize(key)];
-                        } else if (object["is" + capitalize(key)] !== undefined) {
-                            value = object["is" + capitalize(key)];
-                        } else {
-                            value = undefined;
-                        }
-                    }
-
-                    // When resolving an expression we need to pass next_token in case the expression is a function
-                    return Twig.expression.resolveAsync.call(that, value, context, params, next_token, object);
-                })
-                .then(function(result) {
-                    stack.push(result);
-                });
-            }
-        },
-        {
-            type: Twig.expression.type.key.brackets,
-            regex: /^\[([^\]\:]*)\]/,
-            next: Twig.expression.set.operations_extended.concat([
-                    Twig.expression.type.parameter.start]),
-            compile: function(token, stack, output) {
-                var match = token.match[1];
-                delete token.value;
-                delete token.match;
-
-                // The expression stack for the key
-                token.stack = Twig.expression.compile({
-                    value: match
-                }).stack;
-
-                output.push(token);
-            },
-            parse: function(token, stack, context, next_token) {
-                // Evaluate key
-                var that = this,
-                    params = null,
-                    object,
-                    value;
-
-                return parseParams(this, token.params, context)
-                .then(function(parameters) {
-                    params = parameters;
-                    return Twig.expression.parseAsync.call(that, token.stack, context);
-                })
-                .then(function(key) {
-                    object = stack.pop();
-
-                    if (object === null || object === undefined) {
-                        if (that.options.strict_variables) {
-                            throw new Twig.Error("Can't access a key " + key + " on an null or undefined object.");
-                        } else {
-                            return null;
-                        }
-                    }
-
-                    // Get the variable from the context
-                    if (typeof object === 'object' && key in object) {
-                        value = object[key];
-                    } else {
-                        value = null;
-                    }
-
-                    // When resolving an expression we need to pass next_token in case the expression is a function
-                    return Twig.expression.resolveAsync.call(that, value, object, params, next_token);
-                })
-                .then(function(result) {
-                    stack.push(result);
-                });
-            }
-        },
-        {
-            /**
-             * Match a null value.
-             */
-            type: Twig.expression.type._null,
-            // match a number
-            regex: /^(null|NULL|none|NONE)/,
-            next: Twig.expression.set.operations,
-            compile: function(token, stack, output) {
-                delete token.match;
-                token.value = null;
-                output.push(token);
-            },
-            parse: Twig.expression.fn.parse.push_value
-        },
-        {
-            /**
-             * Match the context
-             */
-            type: Twig.expression.type.context,
-            regex: /^_context/,
-            next: Twig.expression.set.operations_extended.concat([
-                    Twig.expression.type.parameter.start]),
-            compile: Twig.expression.fn.compile.push,
-            parse: function(token, stack, context) {
-                stack.push(context);
-            }
-        },
-        {
-            /**
-             * Match a boolean
-             */
-            type: Twig.expression.type.bool,
-            regex: /^(true|TRUE|false|FALSE)/,
-            next: Twig.expression.set.operations,
-            compile: function(token, stack, output) {
-                token.value = (token.match[0].toLowerCase( ) === "true");
-                delete token.match;
-                output.push(token);
-            },
-            parse: Twig.expression.fn.parse.push_value
-        }
-    ];
-
     /**
      * Resolve a context value.
      *
@@ -1071,171 +195,18 @@ module.exports = function (Twig) {
      *                   Should return an object with stack and context set.
      *  }
      *
+     *  Partial token definitions are also allowed, if only property or method
+     *  needs to be overridden or defined.
+     *
      * @param {Object} definition A token definition.
      */
     Twig.expression.extend = function (definition) {
         if (!definition.type) {
             throw new Twig.Error("Unable to extend logic definition. No type provided for " + definition);
         }
-        Twig.expression.handler[definition.type] = definition;
+        Twig.expression.handler[definition.type] = Twig.expression.handler[definition.type] || {};
+        Object.assign(Twig.expression.handler[definition.type], definition);
     };
-
-    // Extend with built-in expressions
-    while (Twig.expression.definitions.length > 0) {
-        Twig.expression.extend(Twig.expression.definitions.shift());
-    }
-
-    /**
-     * Break an expression into tokens defined in Twig.expression.definitions.
-     *
-     * @param {string} expression The string to tokenize.
-     *
-     * @return {Array} An array of tokens.
-     */
-    Twig.expression.tokenize = function (expression) {
-        var tokens = [],
-            // Keep an offset of the location in the expression for error messages.
-            exp_offset = 0,
-            // The valid next tokens of the previous token
-            next = null,
-            // Match information
-            type, regex, regex_i,
-            // The possible next token for the match
-            token_next,
-            // Has a match been found from the definitions
-            match_found, invalid_matches = [], match_function;
-
-        match_function = function () {
-            // Don't pass arguments to `Array.slice`, that is a performance killer
-            var match_i = arguments.length - 2, match = new Array(match_i);
-            while (match_i-- > 0) match[match_i] = arguments[match_i];
-
-            Twig.log.trace("Twig.expression.tokenize",
-                           "Matched a ", type, " regular expression of ", match);
-
-            if (next && Twig.indexOf(next, type) < 0) {
-                invalid_matches.push(
-                    type + " cannot follow a " + tokens[tokens.length - 1].type +
-                           " at template:" + exp_offset + " near '" + match[0].substring(0, 20) +
-                           "...'"
-                );
-
-                // Not a match, don't change the expression
-                return match[0];
-            }
-
-            var handler = Twig.expression.handler[type];
-
-            // Validate the token if a validation function is provided
-            if (handler.validate && !handler.validate(match, tokens)) {
-                return match[0];
-            }
-
-            invalid_matches = [];
-
-            tokens.push({
-                type:  type,
-                value: match[0],
-                match: match
-            });
-
-            match_found = true;
-            next = token_next;
-            exp_offset += match[0].length;
-
-            // Does the token need to return output back to the expression string
-            // e.g. a function match of cycle( might return the '(' back to the expression
-            // This allows look-ahead to differentiate between token types (e.g. functions and variable names)
-            if (handler.transform) {
-                return handler.transform(match, tokens);
-            }
-            return '';
-        };
-
-        Twig.log.debug("Twig.expression.tokenize", "Tokenizing expression ", expression);
-
-        while (expression.length > 0) {
-            expression = expression.trim();
-            for (type in Twig.expression.handler) {
-                token_next = Twig.expression.handler[type].next;
-                regex = Twig.expression.handler[type].regex;
-                Twig.log.trace("Checking type ", type, " on ", expression);
-
-                match_found = false;
-
-                if (Twig.lib.isArray(regex)) {
-                    regex_i = regex.length;
-                    while (regex_i-- > 0)
-                        expression = expression.replace(regex[regex_i], match_function);
-                } else {
-                    expression = expression.replace(regex, match_function);
-                }
-
-                // An expression token has been matched. Break the for loop and start trying to
-                //  match the next template (if expression isn't empty.)
-                if (match_found) {
-                    break;
-                }
-            }
-            if (!match_found) {
-                if (invalid_matches.length > 0) {
-                    throw new Twig.Error(invalid_matches.join(" OR "));
-                } else {
-                    throw new Twig.Error("Unable to parse '" + expression + "' at template position" + exp_offset);
-                }
-            }
-        }
-
-        Twig.log.trace("Twig.expression.tokenize", "Tokenized to ", tokens);
-        return tokens;
-    };
-
-    /**
-     * Compile an expression token.
-     *
-     * @param {Object} raw_token The uncompiled token.
-     *
-     * @return {Object} The compiled token.
-     */
-    Twig.expression.compile = function (raw_token) {
-        var expression = raw_token.value,
-            // Tokenize expression
-            tokens = Twig.expression.tokenize(expression),
-            token = null,
-            output = [],
-            stack = [],
-            token_template = null;
-
-        Twig.log.trace("Twig.expression.compile: ", "Compiling ", expression);
-
-        // Push tokens into RPN stack using the Shunting-yard algorithm
-        // See http://en.wikipedia.org/wiki/Shunting_yard_algorithm
-
-        while (tokens.length > 0) {
-            token = tokens.shift();
-            token_template = Twig.expression.handler[token.type];
-
-            Twig.log.trace("Twig.expression.compile: ", "Compiling ", token);
-
-            // Compile the template
-            token_template.compile && token_template.compile(token, stack, output);
-
-            Twig.log.trace("Twig.expression.compile: ", "Stack is", stack);
-            Twig.log.trace("Twig.expression.compile: ", "Output is", output);
-        }
-
-        while(stack.length > 0) {
-            output.push(stack.pop());
-        }
-
-        Twig.log.trace("Twig.expression.compile: ", "Final output is", output);
-
-        raw_token.stack = output;
-        delete raw_token.value;
-
-        return raw_token;
-    };
-
 
     /**
      * Parse an RPN expression stack within a context.
