+++ src/twig.init.js
@@ -1,0 +1,9 @@
+// ## twig.init.js
+//
+// Defines the initial global Twig object for use in entrypoints.
+
+var Twig = {
+    VERSION: '1.11.0'
+};
+
+module.exports = Twig;
