+++ src/twig.logic.definitions.compile.js
@@ -1,0 +1,619 @@
+// ## twig.logic.compile.js
+//
+// This file handles tokenizing and compiling of logic tokens. {% ... %}
+module.exports = function (Twig) {
+    "use strict";
+
+    // Regular expressions for handling logic tokens.
+    //
+    // Properties:
+    //
+    //      type:  The type of expression this matches
+    //
+    //      regex: A regular expression that matches the format of the token
+    //
+    //      next:  What logic tokens (if any) pop this token off the logic stack. If empty, the
+    //             logic token is assumed to not require an end tag and isn't push onto the stack.
+    //
+    //      open:  Does this tag open a logic expression or is it standalone. For example,
+    //             {% endif %} cannot exist without an opening {% if ... %} tag, so open = false.
+    //
+    //  Functions:
+    //
+    //      compile: A function that handles compiling the token into an output token ready for
+    //               parsing with the parse function.
+    var definitions = [
+        {
+            /**
+             * If type logic tokens.
+             *
+             *  Format: {% if expression %}
+             */
+            type: Twig.logic.type.if_,
+            regex: /^if\s+([\s\S]+)$/,
+            next: [
+                Twig.logic.type.else_,
+                Twig.logic.type.elseif,
+                Twig.logic.type.endif
+            ],
+            open: true,
+            compile: function (token) {
+                var expression = token.match[1];
+                // Compile the expression.
+                token.stack = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * Else if type logic tokens.
+             *
+             *  Format: {% elseif expression %}
+             */
+            type: Twig.logic.type.elseif,
+            regex: /^elseif\s+([^\s].*)$/,
+            next: [
+                Twig.logic.type.else_,
+                Twig.logic.type.elseif,
+                Twig.logic.type.endif
+            ],
+            open: false,
+            compile: function (token) {
+                var expression = token.match[1];
+                // Compile the expression.
+                token.stack = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * Else if type logic tokens.
+             *
+             *  Format: {% elseif expression %}
+             */
+            type: Twig.logic.type.else_,
+            regex: /^else$/,
+            next: [
+                Twig.logic.type.endif,
+                Twig.logic.type.endfor
+            ],
+            open: false
+        },
+        {
+            /**
+             * End if type logic tokens.
+             *
+             *  Format: {% endif %}
+             */
+            type: Twig.logic.type.endif,
+            regex: /^endif$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * For type logic tokens.
+             *
+             *  Format: {% for expression %}
+             */
+            type: Twig.logic.type.for_,
+            regex: /^for\s+([a-zA-Z0-9_,\s]+)\s+in\s+([^\s].*?)(?:\s+if\s+([^\s].*))?$/,
+            next: [
+                Twig.logic.type.else_,
+                Twig.logic.type.endfor
+            ],
+            open: true,
+            compile: function (token) {
+                var key_value = token.match[1],
+                    expression = token.match[2],
+                    conditional = token.match[3],
+                    kv_split = null;
+
+                token.key_var = null;
+                token.value_var = null;
+
+                if (key_value.indexOf(",") >= 0) {
+                    kv_split = key_value.split(',');
+                    if (kv_split.length === 2) {
+                        token.key_var = kv_split[0].trim();
+                        token.value_var = kv_split[1].trim();
+                    } else {
+                        throw new Twig.Error("Invalid expression in for loop: " + key_value);
+                    }
+                } else {
+                    token.value_var = key_value;
+                }
+
+                // Valid expressions for a for loop
+                //   for item     in expression
+                //   for key,item in expression
+
+                // Compile the expression.
+                token.expression = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                // Compile the conditional (if available)
+                if (conditional) {
+                    token.conditional = Twig.expression.compile.call(this, {
+                        type:  Twig.expression.type.expression,
+                        value: conditional
+                    }).stack;
+                }
+
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * End if type logic tokens.
+             *
+             *  Format: {% endif %}
+             */
+            type: Twig.logic.type.endfor,
+            regex: /^endfor$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * Set type logic tokens.
+             *
+             *  Format: {% set key = expression %}
+             */
+            type: Twig.logic.type.set,
+            regex: /^set\s+([a-zA-Z0-9_,\s]+)\s*=\s*([\s\S]+)$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                var key = token.match[1].trim(),
+                    expression = token.match[2],
+                    // Compile the expression.
+                    expression_stack  = Twig.expression.compile.call(this, {
+                        type:  Twig.expression.type.expression,
+                        value: expression
+                    }).stack;
+
+                token.key = key;
+                token.expression = expression_stack;
+
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * Set capture type logic tokens.
+             *
+             *  Format: {% set key %}
+             */
+            type: Twig.logic.type.setcapture,
+            regex: /^set\s+([a-zA-Z0-9_,\s]+)$/,
+            next: [
+                Twig.logic.type.endset
+            ],
+            open: true,
+            compile: function (token) {
+                var key = token.match[1].trim();
+
+                token.key = key;
+
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * End set type block logic tokens.
+             *
+             *  Format: {% endset %}
+             */
+            type: Twig.logic.type.endset,
+            regex: /^endset$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * Filter logic tokens.
+             *
+             *  Format: {% filter upper %} or {% filter lower|escape %}
+             */
+            type: Twig.logic.type.filter,
+            regex: /^filter\s+(.+)$/,
+            next: [
+                Twig.logic.type.endfilter
+            ],
+            open: true,
+            compile: function (token) {
+                var expression = "|" + token.match[1].trim();
+                // Compile the expression.
+                token.stack = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * End filter logic tokens.
+             *
+             *  Format: {% endfilter %}
+             */
+            type: Twig.logic.type.endfilter,
+            regex: /^endfilter$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * Block logic tokens.
+             *
+             *  Format: {% block title %}
+             */
+            type: Twig.logic.type.block,
+            regex: /^block\s+([a-zA-Z0-9_]+)$/,
+            next: [
+                Twig.logic.type.endblock
+            ],
+            open: true,
+            compile: function (token) {
+                token.block = token.match[1].trim();
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * Block shorthand logic tokens.
+             *
+             *  Format: {% block title expression %}
+             */
+            type: Twig.logic.type.shortblock,
+            regex: /^block\s+([a-zA-Z0-9_]+)\s+(.+)$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                token.expression = token.match[2].trim();
+
+                token.output = Twig.expression.compile({
+                    type: Twig.expression.type.expression,
+                    value: token.expression
+                }).stack;
+
+                token.block = token.match[1].trim();
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * End block logic tokens.
+             *
+             *  Format: {% endblock %}
+             */
+            type: Twig.logic.type.endblock,
+            regex: /^endblock(?:\s+([a-zA-Z0-9_]+))?$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * Block logic tokens.
+             *
+             *  Format: {% extends "template.twig" %}
+             */
+            type: Twig.logic.type.extends_,
+            regex: /^extends\s+(.+)$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                var expression = token.match[1].trim();
+                delete token.match;
+
+                token.stack   = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                return token;
+            }
+        },
+        {
+            /**
+             * Block logic tokens.
+             *
+             *  Format: {% use "template.twig" %}
+             */
+            type: Twig.logic.type.use,
+            regex: /^use\s+(.+)$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                var expression = token.match[1].trim();
+                delete token.match;
+
+                token.stack = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                return token;
+            }
+        },
+        {
+            /**
+             * Block logic tokens.
+             *
+             *  Format: {% includes "template.twig" [with {some: 'values'} only] %}
+             */
+            type: Twig.logic.type.include,
+            regex: /^include\s+(.+?)(?:\s|$)(ignore missing(?:\s|$))?(?:with\s+([\S\s]+?))?(?:\s|$)(only)?$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                var match = token.match,
+                    expression = match[1].trim(),
+                    ignoreMissing = match[2] !== undefined,
+                    withContext = match[3],
+                    only = ((match[4] !== undefined) && match[4].length);
+
+                delete token.match;
+
+                token.only = only;
+                token.ignoreMissing = ignoreMissing;
+
+                token.stack = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                if (withContext !== undefined) {
+                    token.withStack = Twig.expression.compile.call(this, {
+                        type:  Twig.expression.type.expression,
+                        value: withContext.trim()
+                    }).stack;
+                }
+
+                return token;
+            }
+        },
+        {
+            type: Twig.logic.type.spaceless,
+            regex: /^spaceless$/,
+            next: [
+                Twig.logic.type.endspaceless
+            ],
+            open: true
+        },
+
+        // Add the {% endspaceless %} token
+        {
+            type: Twig.logic.type.endspaceless,
+            regex: /^endspaceless$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * Macro logic tokens.
+             *
+             * Format: {% maro input(name, value, type, size) %}
+             *
+             */
+            type: Twig.logic.type.macro,
+            regex: /^macro\s+([a-zA-Z0-9_]+)\s*\(\s*((?:[a-zA-Z0-9_]+(?:,\s*)?)*)\s*\)$/,
+            next: [
+                Twig.logic.type.endmacro
+            ],
+            open: true,
+            compile: function (token) {
+                var macroName = token.match[1],
+                    parameters = token.match[2].split(/[\s,]+/);
+
+                //TODO: Clean up duplicate check
+                for (var i=0; i<parameters.length; i++) {
+                    for (var j=0; j<parameters.length; j++){
+                        if (parameters[i] === parameters[j] && i !== j) {
+                            throw new Twig.Error("Duplicate arguments for parameter: "+ parameters[i]);
+                        }
+                    }
+                }
+
+                token.macroName = macroName;
+                token.parameters = parameters;
+
+                delete token.match;
+                return token;
+            }
+        },
+        {
+            /**
+             * End macro logic tokens.
+             *
+             * Format: {% endmacro %}
+             */
+             type: Twig.logic.type.endmacro,
+             regex: /^endmacro$/,
+             next: [ ],
+             open: false
+        },
+        {
+            /*
+            * import logic tokens.
+            *
+            * Format: {% import "template.twig" as form %}
+            */
+            type: Twig.logic.type.import_,
+            regex: /^import\s+(.+)\s+as\s+([a-zA-Z0-9_]+)$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                var expression = token.match[1].trim(),
+                    contextName = token.match[2].trim();
+                delete token.match;
+
+                token.expression = expression;
+                token.contextName = contextName;
+
+                token.stack = Twig.expression.compile.call(this, {
+                    type: Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                return token;
+            }
+        },
+        {
+            /*
+            * from logic tokens.
+            *
+            * Format: {% from "template.twig" import func as form %}
+            */
+            type: Twig.logic.type.from,
+            regex: /^from\s+(.+)\s+import\s+([a-zA-Z0-9_, ]+)$/,
+            next: [ ],
+            open: true,
+            compile: function (token) {
+                var expression = token.match[1].trim(),
+                    macroExpressions = token.match[2].trim().split(/\s*,\s*/),
+                    macroNames = {};
+
+                for (var i=0; i<macroExpressions.length; i++) {
+                    var res = macroExpressions[i];
+
+                    // match function as variable
+                    var macroMatch = res.match(/^([a-zA-Z0-9_]+)\s+as\s+([a-zA-Z0-9_]+)$/);
+                    if (macroMatch) {
+                        macroNames[macroMatch[1].trim()] = macroMatch[2].trim();
+                    }
+                    else if (res.match(/^([a-zA-Z0-9_]+)$/)) {
+                        macroNames[res] = res;
+                    }
+                    else {
+                        // ignore import
+                    }
+
+                }
+
+                delete token.match;
+
+                token.expression = expression;
+                token.macroNames = macroNames;
+
+                token.stack = Twig.expression.compile.call(this, {
+                    type: Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                return token;
+            }
+        },
+        {
+            /**
+             * The embed tag combines the behaviour of include and extends.
+             * It allows you to include another template's contents, just like include does.
+             *
+             *  Format: {% embed "template.twig" [with {some: 'values'} only] %}
+             */
+            type: Twig.logic.type.embed,
+            regex: /^embed\s+(.+?)(?:\s+(ignore missing))?(?:\s+with\s+([\S\s]+?))?(?:\s+(only))?$/,
+            next: [
+                Twig.logic.type.endembed
+            ],
+            open: true,
+            compile: function (token) {
+                var match = token.match,
+                    expression = match[1].trim(),
+                    ignoreMissing = match[2] !== undefined,
+                    withContext = match[3],
+                    only = ((match[4] !== undefined) && match[4].length);
+
+                delete token.match;
+
+                token.only = only;
+                token.ignoreMissing = ignoreMissing;
+
+                token.stack = Twig.expression.compile.call(this, {
+                    type:  Twig.expression.type.expression,
+                    value: expression
+                }).stack;
+
+                if (withContext !== undefined) {
+                    token.withStack = Twig.expression.compile.call(this, {
+                        type:  Twig.expression.type.expression,
+                        value: withContext.trim()
+                    }).stack;
+                }
+
+                return token;
+            }
+        },
+        /* Add the {% endembed %} token
+         *
+         */
+        {
+            type: Twig.logic.type.endembed,
+            regex: /^endembed$/,
+            next: [ ],
+            open: false
+        },
+        {
+            /**
+             * Block logic tokens.
+             *
+             *  Format: {% with {some: 'values'} [only] %}
+             */
+            type: Twig.logic.type['with'],
+            regex: /^(?:with\s+([\S\s]+?))(?:\s|$)(only)?$/,
+            next: [
+                Twig.logic.type.endwith
+            ],
+            open: true,
+            compile: function (token) {
+                var match = token.match,
+                    withContext = match[1],
+                    only = ((match[2] !== undefined) && match[2].length);
+
+                delete token.match;
+
+                token.only = only;
+
+                if (withContext !== undefined) {
+                    token.withStack = Twig.expression.compile.call(this, {
+                        type:  Twig.expression.type.expression,
+                        value: withContext.trim()
+                    }).stack;
+                }
+
+                return token;
+            }
+        },
+        {
+            type: Twig.logic.type.endwith,
+            regex: /^endwith$/,
+            next: [ ],
+            open: false
+        }
+
+    ];
+
+    definitions.forEach(function (definition) {
+        Twig.logic.extend(definition);
+    });
+
+    return Twig;
+
+};
