+++ src/twig.template.compile.js
@@ -1,0 +1,627 @@
+// ## twig.template.compile.js
+//
+// This file handles template loading and compilation.
+module.exports = function (Twig) {
+    "use strict";
+
+    /**
+     * Load a template from a remote location using AJAX and saves in with the given ID.
+     *
+     * Available parameters:
+     *
+     *      async:       Should the HTTP request be performed asynchronously.
+     *                      Defaults to true.
+     *      method:      What method should be used to load the template
+     *                      (fs or ajax)
+     *      parser:      What method should be used to parse the template
+     *                      (twig or source)
+     *      precompiled: Has the template already been compiled.
+     *
+     * @param {string} location  The remote URL to load as a template.
+     * @param {Object} params The template parameters.
+     * @param {function} callback  A callback triggered when the template finishes loading.
+     * @param {function} error_callback  A callback triggered if an error occurs loading the template.
+     *
+     *
+     */
+    Twig.Templates.loadRemote = function(location, params, callback, error_callback) {
+        var loader,
+            // Default to the URL so the template is cached.
+            id = typeof params.id == 'undefined' ? location : params.id,
+            cached = Twig.Templates.registry[id];
+
+        // Check for existing template
+        if (Twig.cache && typeof cached != 'undefined') {
+            // A template is already saved with the given id.
+            if (typeof callback === 'function') {
+                callback(cached);
+            }
+            // TODO: if async, return deferred promise
+            return cached;
+        }
+
+        //if the parser name hasn't been set, default it to twig
+        params.parser = params.parser || 'twig';
+        params.id = id;
+
+        // Default to async
+        if (typeof params.async === 'undefined') {
+            params.async = true;
+        }
+
+        // Assume 'fs' if the loader is not defined
+        loader = this.loaders[params.method] || this.loaders.fs;
+        return loader.call(this, location, params, callback, error_callback);
+    };
+
+    /**
+     * Token syntax definitions.
+     */
+    Twig.token.definitions = [
+        {
+            type: Twig.token.type.raw,
+            open: '{% raw %}',
+            close: '{% endraw %}'
+        },
+        {
+            type: Twig.token.type.raw,
+            open: '{% verbatim %}',
+            close: '{% endverbatim %}'
+        },
+        // *Whitespace type tokens*
+        //
+        // These typically take the form `{{- expression -}}` or `{{- expression }}` or `{{ expression -}}`.
+        {
+            type: Twig.token.type.output_whitespace_pre,
+            open: '{{-',
+            close: '}}'
+        },
+        {
+            type: Twig.token.type.output_whitespace_post,
+            open: '{{',
+            close: '-}}'
+        },
+        {
+            type: Twig.token.type.output_whitespace_both,
+            open: '{{-',
+            close: '-}}'
+        },
+        {
+            type: Twig.token.type.logic_whitespace_pre,
+            open: '{%-',
+            close: '%}'
+        },
+        {
+            type: Twig.token.type.logic_whitespace_post,
+            open: '{%',
+            close: '-%}'
+        },
+        {
+            type: Twig.token.type.logic_whitespace_both,
+            open: '{%-',
+            close: '-%}'
+        },
+        // *Output type tokens*
+        //
+        // These typically take the form `{{ expression }}`.
+        {
+            type: Twig.token.type.output,
+            open: '{{',
+            close: '}}'
+        },
+        // *Logic type tokens*
+        //
+        // These typically take a form like `{% if expression %}` or `{% endif %}`
+        {
+            type: Twig.token.type.logic,
+            open: '{%',
+            close: '%}'
+        },
+        // *Comment type tokens*
+        //
+        // These take the form `{# anything #}`
+        {
+            type: Twig.token.type.comment,
+            open: '{#',
+            close: '#}'
+        }
+    ];
+
+    /**
+     * What characters start "strings" in token definitions. We need this to ignore token close
+     * strings inside an expression.
+     */
+    Twig.token.strings = ['"', "'"];
+
+    Twig.token.findStart = function (template) {
+        var output = {
+                position: null,
+                def: null
+            },
+            close_position = null,
+            len = Twig.token.definitions.length,
+            i,
+            token_template,
+            first_key_position,
+            close_key_position;
+
+        for (i=0;i<len;i++) {
+            token_template = Twig.token.definitions[i];
+            first_key_position = template.indexOf(token_template.open);
+            close_key_position = template.indexOf(token_template.close);
+
+            Twig.log.trace("Twig.token.findStart: ", "Searching for ", token_template.open, " found at ", first_key_position);
+
+            //Special handling for mismatched tokens
+            if (first_key_position >= 0) {
+                //This token matches the template
+                if (token_template.open.length !== token_template.close.length) {
+                    //This token has mismatched closing and opening tags
+                    if (close_key_position < 0) {
+                        //This token's closing tag does not match the template
+                        continue;
+                    }
+                }
+            }
+            // Does this token occur before any other types?
+            if (first_key_position >= 0 && (output.position === null || first_key_position < output.position)) {
+                output.position = first_key_position;
+                output.def = token_template;
+                close_position = close_key_position;
+            } else if (first_key_position >= 0 && output.position !== null && first_key_position === output.position) {
+                /*This token exactly matches another token,
+                 greedily match to check if this token has a greater specificity*/
+                if (token_template.open.length > output.def.open.length) {
+                    //This token's opening tag is more specific than the previous match
+                    output.position = first_key_position;
+                    output.def = token_template;
+                    close_position = close_key_position;
+                } else if (token_template.open.length === output.def.open.length) {
+                    if (token_template.close.length > output.def.close.length) {
+                        //This token's opening tag is as specific as the previous match,
+                        //but the closing tag has greater specificity
+                        if (close_key_position >= 0 && close_key_position < close_position) {
+                            //This token's closing tag exists in the template,
+                            //and it occurs sooner than the previous match
+                            output.position = first_key_position;
+                            output.def = token_template;
+                            close_position = close_key_position;
+                        }
+                    } else if (close_key_position >= 0 && close_key_position < close_position) {
+                        //This token's closing tag is not more specific than the previous match,
+                        //but it occurs sooner than the previous match
+                        output.position = first_key_position;
+                        output.def = token_template;
+                        close_position = close_key_position;
+                    }
+                }
+            }
+        }
+
+        // delete output['close_position'];
+
+        return output;
+    };
+
+    Twig.token.findEnd = function (template, token_def, start) {
+        var end = null,
+            found = false,
+            offset = 0,
+
+            // String position variables
+            str_pos = null,
+            str_found = null,
+            pos = null,
+            end_offset = null,
+            this_str_pos = null,
+            end_str_pos = null,
+
+            // For loop variables
+            i,
+            l;
+
+        while (!found) {
+            str_pos = null;
+            str_found = null;
+            pos = template.indexOf(token_def.close, offset);
+
+            if (pos >= 0) {
+                end = pos;
+                found = true;
+            } else {
+                // throw an exception
+                throw new Twig.Error("Unable to find closing bracket '" + token_def.close +
+                    "'" + " opened near template position " + start);
+            }
+
+            // Ignore quotes within comments; just look for the next comment close sequence,
+            // regardless of what comes before it. https://github.com/justjohn/twig.js/issues/95
+            if (token_def.type === Twig.token.type.comment) {
+                break;
+            }
+            // Ignore quotes within raw tag
+            // Fixes #283
+            if (token_def.type === Twig.token.type.raw) {
+                break;
+            }
+
+            l = Twig.token.strings.length;
+            for (i = 0; i < l; i += 1) {
+                this_str_pos = template.indexOf(Twig.token.strings[i], offset);
+
+                if (this_str_pos > 0 && this_str_pos < pos &&
+                    (str_pos === null || this_str_pos < str_pos)) {
+                    str_pos = this_str_pos;
+                    str_found = Twig.token.strings[i];
+                }
+            }
+
+            // We found a string before the end of the token, now find the string's end and set the search offset to it
+            if (str_pos !== null) {
+                end_offset = str_pos + 1;
+                end = null;
+                found = false;
+                while (true) {
+                    end_str_pos = template.indexOf(str_found, end_offset);
+                    if (end_str_pos < 0) {
+                        throw "Unclosed string in template";
+                    }
+                    // Ignore escaped quotes
+                    if (template.substr(end_str_pos - 1, 1) !== "\\") {
+                        offset = end_str_pos + 1;
+                        break;
+                    } else {
+                        end_offset = end_str_pos + 1;
+                    }
+                }
+            }
+        }
+        return end;
+    };
+
+    Twig.compile = function (tokens) {
+        var self = this;
+        return Twig.attempt(function() {
+
+            // Output and intermediate stacks
+            var output = [],
+                stack = [],
+                // The tokens between open and close tags
+                intermediate_output = [],
+
+                token = null,
+                logic_token = null,
+                unclosed_token = null,
+                // Temporary previous token.
+                prev_token = null,
+                // Temporary previous output.
+                prev_output = null,
+                // Temporary previous intermediate output.
+                prev_intermediate_output = null,
+                // The previous token's template
+                prev_template = null,
+                // Token lookahead
+                next_token = null,
+                // The output token
+                tok_output = null,
+
+                // Logic Token values
+                type = null,
+                open = null,
+                next = null;
+
+            var compile_output = function(token) {
+                Twig.expression.compile.call(self, token);
+                if (stack.length > 0) {
+                    intermediate_output.push(token);
+                } else {
+                    output.push(token);
+                }
+            };
+
+            var compile_logic = function(token) {
+                // Compile the logic token
+                logic_token = Twig.logic.compile.call(self, token);
+
+                type = logic_token.type;
+                open = Twig.logic.handler[type].open;
+                next = Twig.logic.handler[type].next;
+
+                Twig.log.trace("Twig.compile: ", "Compiled logic token to ", logic_token,
+                    " next is: ", next, " open is : ", open);
+
+                // Not a standalone token, check logic stack to see if this is expected
+                if (open !== undefined && !open) {
+                    prev_token = stack.pop();
+                    prev_template = Twig.logic.handler[prev_token.type];
+
+                    if (Twig.indexOf(prev_template.next, type) < 0) {
+                        throw new Error(type + " not expected after a " + prev_token.type);
+                    }
+
+                    prev_token.output = prev_token.output || [];
+
+                    prev_token.output = prev_token.output.concat(intermediate_output);
+                    intermediate_output = [];
+
+                    tok_output = {
+                        type: Twig.token.type.logic,
+                        token: prev_token
+                    };
+                    if (stack.length > 0) {
+                        intermediate_output.push(tok_output);
+                    } else {
+                        output.push(tok_output);
+                    }
+                }
+
+                // This token requires additional tokens to complete the logic structure.
+                if (next !== undefined && next.length > 0) {
+                    Twig.log.trace("Twig.compile: ", "Pushing ", logic_token, " to logic stack.");
+
+                    if (stack.length > 0) {
+                        // Put any currently held output into the output list of the logic operator
+                        // currently at the head of the stack before we push a new one on.
+                        prev_token = stack.pop();
+                        prev_token.output = prev_token.output || [];
+                        prev_token.output = prev_token.output.concat(intermediate_output);
+                        stack.push(prev_token);
+                        intermediate_output = [];
+                    }
+
+                    // Push the new logic token onto the logic stack
+                    stack.push(logic_token);
+
+                } else if (open !== undefined && open) {
+                    tok_output = {
+                        type: Twig.token.type.logic,
+                        token: logic_token
+                    };
+                    // Standalone token (like {% set ... %}
+                    if (stack.length > 0) {
+                        intermediate_output.push(tok_output);
+                    } else {
+                        output.push(tok_output);
+                    }
+                }
+            };
+
+            while (tokens.length > 0) {
+                token = tokens.shift();
+                prev_output = output[output.length - 1];
+                prev_intermediate_output = intermediate_output[intermediate_output.length - 1];
+                next_token = tokens[0];
+                Twig.log.trace("Compiling token ", token);
+                switch (token.type) {
+                    case Twig.token.type.raw:
+                        if (stack.length > 0) {
+                            intermediate_output.push(token);
+                        } else {
+                            output.push(token);
+                        }
+                        break;
+
+                    case Twig.token.type.logic:
+                        compile_logic.call(self, token);
+                        break;
+
+                    // Do nothing, comments should be ignored
+                    case Twig.token.type.comment:
+                        break;
+
+                    case Twig.token.type.output:
+                        compile_output.call(self, token);
+                        break;
+
+                    //Kill whitespace ahead and behind this token
+                    case Twig.token.type.logic_whitespace_pre:
+                    case Twig.token.type.logic_whitespace_post:
+                    case Twig.token.type.logic_whitespace_both:
+                    case Twig.token.type.output_whitespace_pre:
+                    case Twig.token.type.output_whitespace_post:
+                    case Twig.token.type.output_whitespace_both:
+                        if (token.type !== Twig.token.type.output_whitespace_post && token.type !== Twig.token.type.logic_whitespace_post) {
+                            if (prev_output) {
+                                //If the previous output is raw, pop it off
+                                if (prev_output.type === Twig.token.type.raw) {
+                                    output.pop();
+
+                                    //If the previous output is not just whitespace, trim it
+                                    if (prev_output.value.match(/^\s*$/) === null) {
+                                        prev_output.value = prev_output.value.trim();
+                                        //Repush the previous output
+                                        output.push(prev_output);
+                                    }
+                                }
+                            }
+
+                            if (prev_intermediate_output) {
+                                //If the previous intermediate output is raw, pop it off
+                                if (prev_intermediate_output.type === Twig.token.type.raw) {
+                                    intermediate_output.pop();
+
+                                    //If the previous output is not just whitespace, trim it
+                                    if (prev_intermediate_output.value.match(/^\s*$/) === null) {
+                                        prev_intermediate_output.value = prev_intermediate_output.value.trim();
+                                        //Repush the previous intermediate output
+                                        intermediate_output.push(prev_intermediate_output);
+                                    }
+                                }
+                            }
+                        }
+
+                        //Compile this token
+                        switch (token.type) {
+                            case Twig.token.type.output_whitespace_pre:
+                            case Twig.token.type.output_whitespace_post:
+                            case Twig.token.type.output_whitespace_both:
+                                compile_output.call(self, token);
+                                break;
+                            case Twig.token.type.logic_whitespace_pre:
+                            case Twig.token.type.logic_whitespace_post:
+                            case Twig.token.type.logic_whitespace_both:
+                                compile_logic.call(self, token);
+                                break;
+                        }
+
+                        if (token.type !== Twig.token.type.output_whitespace_pre && token.type !== Twig.token.type.logic_whitespace_pre) {
+                            if (next_token) {
+                                //If the next token is raw, shift it out
+                                if (next_token.type === Twig.token.type.raw) {
+                                    tokens.shift();
+
+                                    //If the next token is not just whitespace, trim it
+                                    if (next_token.value.match(/^\s*$/) === null) {
+                                        next_token.value = next_token.value.trim();
+                                        //Unshift the next token
+                                        tokens.unshift(next_token);
+                                    }
+                                }
+                            }
+                        }
+
+                        break;
+                }
+
+                Twig.log.trace("Twig.compile: ", " Output: ", output,
+                    " Logic Stack: ", stack,
+                    " Pending Output: ", intermediate_output );
+            }
+
+            // Verify that there are no logic tokens left in the stack.
+            if (stack.length > 0) {
+                unclosed_token = stack.pop();
+                throw new Error("Unable to find an end tag for " + unclosed_token.type +
+                    ", expecting one of " + unclosed_token.next);
+            }
+            return output;
+        }, function(ex) {
+            if (self.options.rethrow) {
+                if (ex.type == 'TwigException' && !ex.file) {
+                    ex.file = self.id;
+                }
+
+                throw ex
+            }
+            else {
+                Twig.log.error("Error compiling twig template " + self.id + ": ");
+                if (ex.stack) {
+                    Twig.log.error(ex.stack);
+                } else {
+                    Twig.log.error(ex.toString());
+                }
+            }
+        });
+    };
+
+    /**
+     * Convert a template into high-level tokens.
+     */
+    Twig.tokenize = function (template) {
+        var tokens = [],
+            // An offset for reporting errors locations in the template.
+            error_offset = 0,
+
+            // The start and type of the first token found in the template.
+            found_token = null,
+            // The end position of the matched token.
+            end = null;
+
+        while (template.length > 0) {
+            // Find the first occurance of any token type in the template
+            found_token = Twig.token.findStart(template);
+
+            Twig.log.trace("Twig.tokenize: ", "Found token: ", found_token);
+
+            if (found_token.position !== null) {
+                // Add a raw type token for anything before the start of the token
+                if (found_token.position > 0) {
+                    tokens.push({
+                        type: Twig.token.type.raw,
+                        value: template.substring(0, found_token.position)
+                    });
+                }
+                template = template.substr(found_token.position + found_token.def.open.length);
+                error_offset += found_token.position + found_token.def.open.length;
+
+                // Find the end of the token
+                end = Twig.token.findEnd(template, found_token.def, error_offset);
+
+                Twig.log.trace("Twig.tokenize: ", "Token ends at ", end);
+
+                tokens.push({
+                    type:  found_token.def.type,
+                    value: template.substring(0, end).trim()
+                });
+
+                if (template.substr( end + found_token.def.close.length, 1 ) === "\n") {
+                    switch (found_token.def.type) {
+                        case "logic_whitespace_pre":
+                        case "logic_whitespace_post":
+                        case "logic_whitespace_both":
+                        case "logic":
+                            // Newlines directly after logic tokens are ignored
+                            end += 1;
+                            break;
+                    }
+                }
+
+                template = template.substr(end + found_token.def.close.length);
+
+                // Increment the position in the template
+                error_offset += end + found_token.def.close.length;
+
+            } else {
+                // No more tokens -> add the rest of the template as a raw-type token
+                tokens.push({
+                    type: Twig.token.type.raw,
+                    value: template
+                });
+                template = '';
+            }
+        }
+
+        return tokens;
+    };
+
+    Twig.Template.prototype.getLoaderMethod = function() {
+        if (this.path) {
+            return 'fs';
+        }
+        if (this.url) {
+            return 'ajax';
+        }
+        return this.method || 'fs';
+    };
+
+    Twig.Template.prototype.compile = function(options) {
+        // compile the template into raw JS
+        return Twig.compiler.compile(this, options);
+    };
+
+    /**
+     * Tokenize and compile a string template.
+     *
+     * @param {string} data The template.
+     *
+     * @return {Array} The compiled tokens.
+     */
+    Twig.prepare = function(data) {
+        var tokens, raw_tokens;
+
+        // Tokenize
+        Twig.log.debug("Twig.prepare: ", "Tokenizing ", data);
+        raw_tokens = Twig.tokenize.call(this, data);
+
+        // Compile
+        Twig.log.debug("Twig.prepare: ", "Compiling ", raw_tokens);
+        tokens = Twig.compile.call(this, raw_tokens);
+
+        Twig.log.debug("Twig.prepare: ", "Compiled ", tokens);
+
+        return tokens;
+    };
+
+    return Twig;
+
+};
