+++ src/twig.template.js
@@ -1,0 +1,410 @@
+// ## twig.template.js
+//
+// This file handles template parsing.
+module.exports = function (Twig) {
+    "use strict";
+
+    // Namespace for template storage and retrieval
+    Twig.Templates = {
+        /**
+         * Registered template loaders - use Twig.Templates.registerLoader to add supported loaders
+         * @type {Object}
+         */
+        loaders: {},
+
+        /**
+         * Registered template parsers - use Twig.Templates.registerParser to add supported parsers
+         * @type {Object}
+         */
+        parsers: {},
+
+        /**
+         * Cached / loaded templates
+         * @type {Object}
+         */
+        registry: {}
+    };
+
+    /**
+     * Is this id valid for a twig template?
+     *
+     * @param {string} id The ID to check.
+     *
+     * @throws {Twig.Error} If the ID is invalid or used.
+     * @return {boolean} True if the ID is valid.
+     */
+    Twig.validateId = function(id) {
+        if (id === "prototype") {
+            throw new Twig.Error(id + " is not a valid twig identifier");
+        } else if (Twig.cache && Twig.Templates.registry.hasOwnProperty(id)) {
+            throw new Twig.Error("There is already a template with the ID " + id);
+        }
+        return true;
+    };
+
+    /**
+     * Register a template loader
+     *
+     * @example
+     * Twig.extend(function(Twig) {
+     *    Twig.Templates.registerLoader('custom_loader', function(location, params, callback, error_callback) {
+     *        // ... load the template ...
+     *        params.data = loadedTemplateData;
+     *        // create and return the template
+     *        var template = new Twig.Template(params);
+     *        if (typeof callback === 'function') {
+     *            callback(template);
+     *        }
+     *        return template;
+     *    });
+     * });
+     *
+     * @param {String} method_name The method this loader is intended for (ajax, fs)
+     * @param {Function} func The function to execute when loading the template
+     * @param {Object|undefined} scope Optional scope parameter to bind func to
+     *
+     * @throws Twig.Error
+     *
+     * @return {void}
+     */
+    Twig.Templates.registerLoader = function(method_name, func, scope) {
+        if (typeof func !== 'function') {
+            throw new Twig.Error('Unable to add loader for ' + method_name + ': Invalid function reference given.');
+        }
+        if (scope) {
+            func = func.bind(scope);
+        }
+        this.loaders[method_name] = func;
+    };
+
+    /**
+     * Remove a registered loader
+     *
+     * @param {String} method_name The method name for the loader you wish to remove
+     *
+     * @return {void}
+     */
+    Twig.Templates.unRegisterLoader = function(method_name) {
+        if (this.isRegisteredLoader(method_name)) {
+            delete this.loaders[method_name];
+        }
+    };
+
+    /**
+     * See if a loader is registered by its method name
+     *
+     * @param {String} method_name The name of the loader you are looking for
+     *
+     * @return {boolean}
+     */
+    Twig.Templates.isRegisteredLoader = function(method_name) {
+        return this.loaders.hasOwnProperty(method_name);
+    };
+
+    /**
+     * Register a template parser
+     *
+     * @example
+     * Twig.extend(function(Twig) {
+     *    Twig.Templates.registerParser('custom_parser', function(params) {
+     *        // this template source can be accessed in params.data
+     *        var template = params.data
+     *
+     *        // ... custom process that modifies the template
+     *
+     *        // return the parsed template
+     *        return template;
+     *    });
+     * });
+     *
+     * @param {String} method_name The method this parser is intended for (twig, source)
+     * @param {Function} func The function to execute when parsing the template
+     * @param {Object|undefined} scope Optional scope parameter to bind func to
+     *
+     * @throws Twig.Error
+     *
+     * @return {void}
+     */
+    Twig.Templates.registerParser = function(method_name, func, scope) {
+        if (typeof func !== 'function') {
+            throw new Twig.Error('Unable to add parser for ' + method_name + ': Invalid function regerence given.');
+        }
+
+        if (scope) {
+            func = func.bind(scope);
+        }
+
+        this.parsers[method_name] = func;
+    };
+
+    /**
+     * Remove a registered parser
+     *
+     * @param {String} method_name The method name for the parser you wish to remove
+     *
+     * @return {void}
+     */
+    Twig.Templates.unRegisterParser = function(method_name) {
+        if (this.isRegisteredParser(method_name)) {
+            delete this.parsers[method_name];
+        }
+    };
+
+    /**
+     * See if a parser is registered by its method name
+     *
+     * @param {String} method_name The name of the parser you are looking for
+     *
+     * @return {boolean}
+     */
+    Twig.Templates.isRegisteredParser = function(method_name) {
+        return this.parsers.hasOwnProperty(method_name);
+    };
+
+    /**
+     * Save a template object to the store.
+     *
+     * @param {Twig.Template} template   The twig.js template to store.
+     */
+    Twig.Templates.save = function(template) {
+        if (template.id === undefined) {
+            throw new Twig.Error("Unable to save template with no id");
+        }
+        Twig.Templates.registry[template.id] = template;
+    };
+
+    /**
+     * Load a previously saved template from the store.
+     *
+     * @param {string} id   The ID of the template to load.
+     *
+     * @return {Twig.Template} A twig.js template stored with the provided ID.
+     */
+    Twig.Templates.load = function(id) {
+        if (!Twig.Templates.registry.hasOwnProperty(id)) {
+            return null;
+        }
+        return Twig.Templates.registry[id];
+    };
+
+    // Determine object type
+    function is(type, obj) {
+        var clas = Object.prototype.toString.call(obj).slice(8, -1);
+        return obj !== undefined && obj !== null && clas === type;
+    }
+
+    /**
+     * Create a new twig.js template.
+     *
+     * Parameters: {
+     *      data:   The template, either pre-compiled tokens or a string template
+     *      id:     The name of this template
+     *      blocks: Any pre-existing block from a child template
+     * }
+     *
+     * @param {Object} params The template parameters.
+     */
+    Twig.Template = function ( params ) {
+        var data = params.data,
+            id = params.id,
+            blocks = params.blocks,
+            macros = params.macros || {},
+            base = params.base,
+            path = params.path,
+            url = params.url,
+            name = params.name,
+            method = params.method,
+            // parser options
+            options = params.options;
+
+        // # What is stored in a Twig.Template
+        //
+        // The Twig Template hold several chucks of data.
+        //
+        //     {
+        //          id:     The token ID (if any)
+        //          tokens: The list of tokens that makes up this template.
+        //          blocks: The list of block this template contains.
+        //          base:   The base template (if any)
+        //            options:  {
+        //                Compiler/parser options
+        //
+        //                strict_variables: true/false
+        //                    Should missing variable/keys emit an error message. If false, they default to null.
+        //            }
+        //     }
+        //
+
+        this.id     = id;
+        this.method = method;
+        this.base   = base;
+        this.path   = path;
+        this.url    = url;
+        this.name   = name;
+        this.macros = macros;
+        this.options = options;
+
+        this.reset(blocks);
+
+        if (is('String', data)) {
+            this.tokens = Twig.prepare.call(this, data);
+        } else {
+            this.tokens = data;
+        }
+
+        if (id !== undefined) {
+            Twig.Templates.save(this);
+        }
+    };
+
+    Twig.Template.prototype.reset = function(blocks) {
+        Twig.log.debug("Twig.Template.reset", "Reseting template " + this.id);
+        this.blocks = {};
+        this.importedBlocks = [];
+        this.originalBlockTokens = {};
+        this.child = {
+            blocks: blocks || {}
+        };
+        this.extend = null;
+        this.parseStack = [];
+    };
+
+    Twig.Template.prototype.render = function (context, params, allow_async) {
+        var that = this;
+
+        this.context = context || {};
+
+        // Clear any previous state
+        this.reset();
+        if (params && params.blocks) {
+            this.blocks = params.blocks;
+        }
+        if (params && params.macros) {
+            this.macros = params.macros;
+        }
+
+        return Twig.async.potentiallyAsync(this, allow_async, function() {
+            return Twig.parseAsync.call(this, this.tokens, this.context)
+            .then(function(output) {
+                var ext_template,
+                    url;
+
+                // Does this template extend another
+                if (that.extend) {
+
+                    // check if the template is provided inline
+                    if ( that.options.allowInlineIncludes ) {
+                        ext_template = Twig.Templates.load(that.extend);
+                        if ( ext_template ) {
+                            ext_template.options = that.options;
+                        }
+                    }
+
+                    // check for the template file via include
+                    if (!ext_template) {
+                        url = Twig.path.parsePath(that, that.extend);
+
+                        ext_template = Twig.Templates.loadRemote(url, {
+                            method: that.getLoaderMethod(),
+                            base: that.base,
+                            async:  false,
+                            id:     url,
+                            options: that.options
+                        });
+                    }
+
+                    that.parent = ext_template;
+
+                    return that.parent.renderAsync(that.context, {
+                        blocks: that.blocks
+                    });
+                }
+
+                if (!params) {
+                    return output;
+                } else if (params.output == 'blocks') {
+                    return that.blocks;
+                } else if (params.output == 'macros') {
+                    return that.macros;
+                } else {
+                    return output;
+                }
+            });
+        });
+    };
+
+    Twig.Template.prototype.importFile = function(file) {
+        var url, sub_template;
+        if (!this.url && this.options.allowInlineIncludes) {
+            file = this.path ? Twig.path.parsePath(this, file) : file;
+            sub_template = Twig.Templates.load(file);
+
+            if (!sub_template) {
+                sub_template = Twig.Templates.loadRemote(url, {
+                    id: file,
+                    method: this.getLoaderMethod(),
+                    async: false,
+                    path: file,
+                    options: this.options
+                });
+
+                if (!sub_template) {
+                    throw new Twig.Error("Unable to find the template " + file);
+                }
+            }
+
+            sub_template.options = this.options;
+
+            return sub_template;
+        }
+
+        url = Twig.path.parsePath(this, file);
+
+        // Load blocks from an external file
+        sub_template = Twig.Templates.loadRemote(url, {
+            method: this.getLoaderMethod(),
+            base: this.base,
+            async: false,
+            options: this.options,
+            id: url
+        });
+
+        return sub_template;
+    };
+
+
+    Twig.Template.prototype.importBlocks = function(file, override) {
+        var sub_template = this.importFile(file),
+            context = this.context,
+            that = this,
+            key;
+
+        override = override || false;
+
+        sub_template.render(context);
+
+        // Mixin blocks
+        Twig.forEach(Object.keys(sub_template.blocks), function(key) {
+            if (override || that.blocks[key] === undefined) {
+                that.blocks[key] = sub_template.blocks[key];
+                that.importedBlocks.push(key);
+            }
+        });
+    };
+
+    Twig.Template.prototype.importMacros = function(file) {
+        var url = Twig.path.parsePath(this, file);
+
+        // load remote template
+        var remoteTemplate = Twig.Templates.loadRemote(url, {
+            method: this.getLoaderMethod(),
+            async: false,
+            id: url
+        });
+
+        return remoteTemplate;
+    };
+
+    return Twig;
+
+};
